Chalmers SysBio GEM in the form of gene-metabolite interactions

Processing GEMs from Wang et al., 2021 (https://github.com/SysBioChalmers) to generate PKN for COSMOS

Usage

chalmers_gem_network(
  organism_or_gem = "Human",
  metab_max_degree = 400L,
  protein_ids = c("uniprot", "genesymbol"),
  metabolite_ids = c("hmdb", "kegg")
)

Arguments

organism_or_gem

Character or integer or list or data frame: either an organism (taxon) identifier or a list containing the “reactions“ data frame as it is provided by chalmers_gem, or the reactions data frame itself. Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus), 10116 (Rattus norvegicus), 7955 (Danio rerio), 7227 (Drosophila melanogaster) and 6239 (Caenorhabditis elegans).

metab_max_degree

Degree cutoff used to prune metabolites with high degree assuming they are cofactors (400 by default).

protein_ids

Character: translate the protein identifiers to these ID types. Each ID type results two extra columns in the output, for the "a" and "b" sides of the interaction, respectively. The default ID type for proteins is Esembl Gene ID, and by default UniProt IDs and Gene Symbols are included.

metabolite_ids

Character: translate the protein identifiers to these ID types. Each ID type results two extra columns in the output, for the "a" and "b" sides of the interaction, respectively. The default ID type for metabolites is Metabolic Atlas ID, and HMDB IDs and KEGG IDs are included.

Value

Data frame (tibble) of gene-metabolite interactions.

References

Wang H, Robinson JL, Kocabas P, Gustafsson J, Anton M, Cholley PE, Huang S, Gobom J, Svensson T, Uhlen M, Zetterberg H, Nielsen J. Genome-scale metabolic network reconstruction of model animals as a platform for translational research. Proc Natl Acad Sci U S A. 2021 Jul 27;118(30):e2102344118. doi: doi:10.1073/pnas.2102344118 .

See also

Examples

gem <- chalmers_gem_network()
#> Error in map2(.x, .y, .f, ..., .progress = .progress): ℹ In index: 1.
#> ℹ With name: col_types.
#> Caused by error in `loadNamespace()`:
#> ! there is no package called ‘R.matlab’