GEM matlab file from Chalmers Sysbio (Wang et al., 2021)

Downloads and imports the matlab file containing the genome scale metabolic models created by Chalmers SysBio.

Usage

chalmers_gem_raw(organism = "Human")

Arguments

organism

Character or integer: name or identifier of the organism. Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus), 10116 (Rattus norvegicu), 7955 (Danio rerio), 7227 (Drosophila melanogaster) and 6239 (Caenorhabditis elegans).

Value

Matlab object containing the GEM.

Details

The Matlab object is parsed into a nested list containing a number of vectors and two sparse matrices. The top level contains a single element under the name "ihuman" for human; under this key there is an array of 31 elements. These elements are labeled by the row names of the array.

References

Wang H, Robinson JL, Kocabas P, Gustafsson J, Anton M, Cholley PE, Huang S, Gobom J, Svensson T, Uhlen M, Zetterberg H, Nielsen J. Genome-scale metabolic network reconstruction of model animals as a platform for translational research. Proc Natl Acad Sci U S A. 2021 Jul 27;118(30):e2102344118. doi: doi:10.1073/pnas.2102344118 .

See also

Examples

chalmers_gem_raw()
#> $ihuman
#> , , 1
#> 
#>                     [,1]                                                  
#> id                  "Human-GEM"                                           
#> description         "Generic genome-scale metabolic model of Homo sapiens"
#> rxns                list,12995                                            
#> mets                list,8456                                             
#> S                   <S4 class ‘dgCMatrix’ [package “Matrix”] with 6 slots>
#> lb                  numeric,12995                                         
#> ub                  numeric,12995                                         
#> rev                 numeric,12995                                         
#> c                   numeric,12995                                         
#> b                   numeric,8456                                          
#> comps               list,9                                                
#> compNames           list,9                                                
#> rxnNames            list,12995                                            
#> grRules             list,12995                                            
#> rxnGeneMat          <S4 class ‘dgCMatrix’ [package “Matrix”] with 6 slots>
#> subSystems          list,12995                                            
#> eccodes             list,12995                                            
#> rxnNotes            list,12995                                            
#> genes               list,2889                                             
#> geneShortNames      list,2889                                             
#> metNames            list,8456                                             
#> metComps            numeric,8456                                          
#> inchis              list,8456                                             
#> metFormulas         list,8456                                             
#> rxnReferences       list,12995                                            
#> rxnFrom             list,12995                                            
#> metFrom             list,8456                                             
#> rxnConfidenceScores numeric,12995                                         
#> metCharges          numeric,8456                                          
#> version             "1.18.0"                                              
#> annotation          list,5                                                
#> 
#> 
#> attr(,"header")
#> attr(,"header")$description
#> [1] "MATLAB 5.0 MAT-file, Platform: MACI64, Created on: Fri Dec  8 14:17:06 2023                                         "
#> 
#> attr(,"header")$version
#> [1] "5"
#> 
#> attr(,"header")$endian
#> [1] "little"
#> 
#> attr(,"origin")
#> [1] "cache"
#> attr(,"resource")
#> [1] "Chalmers GEM"
#> attr(,"url")
#> [1] "https://raw.githubusercontent.com/SysBioChalmers/Human-GEM/main/model/Human-GEM.mat"
#> attr(,"source")
#> [1] "Chalmers GEM (raw.githubusercontent.com)"