TF-target interactions from CollecTRI

CollecTRI is a comprehensive resource of transcriptional regulation, published in 2023, consisting of 14 resources and original literature curation.

Usage

collectri(
  resources = NULL,
  organism = 9606L,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = TRUE,
  ...
)

Arguments

resources

interactions not reported in these databases are removed. See get_interaction_resources for more information.

organism

Interactions are available for human, mouse and rat. Choose among: 9606 human (default), 10116 rat and 10090 Mouse

references_by_resource

if FALSE, removes the resource name prefixes from the references (PubMed IDs); this way the information which reference comes from which resource will be lost and the PubMed IDs will be unique.

exclude

Character: datasets or resources to exclude.

strict_evidences

Logical: restrict the evidences to the queried datasets and resources. If set to FALSE, the directions and effect signs and references might be based on other datasets and resources. In case of CollecTRI this is not desirable for most of the applications. For most of the other interaction querying functions it is `FALSE` by default.

...

Optional additional arguments, passed to import_transcriptional_interactions.

Value

A dataframe of TF-target interactions.

See also

Examples

collectri_grn <- collectri()
#> Error: The output colorization was canceled. isatty(stdout()) returned FALSE.
collectri_grn
#> Error in eval(expr, envir, enclos): object 'collectri_grn' not found