Curated ligand-receptor interactions

The OmniPath intercell database annotates individual proteins and complexes, and we combine these annotations with network interactions on the client side, using import_intercell_network. The architecture of this database is complex, aiming to cover a broad range of knowledge on various levels of details and confidence. We can use the filter_intercell and filter_intercell_network functions for automated, data driven quality filtering, in order to enrich the cell-cell communication network in higher confidence interactions. However, for many users, a simple combination of the most established, expert curated ligand-receptor resources, provided by this function, fits better their purpose.

curated_ligand_receptor_interactions(
  curated_resources = c("Guide2Pharma", "HPMR", "ICELLNET", "CellTalkDB", "CellChatDB",
    "connectomeDB2020"),
  cellphonedb = TRUE,
  signalink = TRUE
)

Arguments

curated_resources

Character vector of the resource names which are considered to be expert curated. You can include any post-translational network resource here, but if you include non ligand-receptor or non curated resources, the result will not fulfill the original intention of this function.

cellphonedb

Logical: include the curated interactions from CellPhoneDB (not the whole CellPhoneDB but a subset of it).

signalink

Logical: include the ligand-receptor interactions from SignaLink. These are all expert curated.

Value

A data frame similar to interactions (network) data frames, the source and target partners being ligand and receptor, respectively.

Details

Some resources are a mixture of curated and bulk imported interactions, and sometimes it's not trivial to separate these, we take care of these here. This function does not use the intercell database of OmniPath, but retrieves and filters a handful of network resources. The returned data frame has the layout of interactions (network) data frames, and the source and target partners implicitly correspond to ligand and receptor. The data frame shows all resources and references for all interactions, but each interaction is supported by at least one ligand-receptor resource which is supposed to based on expert curation in a ligand-receptor context.

See also

Examples

lr <- curated_ligand_receptor_interactions()
#> Error: The output colorization was canceled. isatty(stdout()) returned FALSE.
lr
#> # A tibble: 5,292 x 15
#>    source target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>  <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 P0DP23 Q13507 CALM1             TRPC3                       1              0
#>  2 P60903 Q9H1D0 S100A10           TRPV6                       1              1
#>  3 Q9Y219 P46531 JAG2              NOTCH1                      1              1
#>  4 O00548 P46531 DLL1              NOTCH1                      1              1
#>  5 P05019 P08069 IGF1              IGF1R                       1              1
#>  6 P08069 P05019 IGF1R             IGF1                        1              0
#>  7 P46531 P78504 NOTCH1            JAG1                        1              1
#>  8 P78504 P46531 JAG1              NOTCH1                      1              1
#>  9 P41221 Q14332 WNT5A             FZD2                        1              1
#> 10 Q9GZX6 Q08334 IL22              IL10RB                      1              1
#> # … with 5,282 more rows, and 9 more variables: is_inhibition <dbl>,
#> #   consensus_direction <dbl>, consensus_stimulation <dbl>,
#> #   consensus_inhibition <dbl>, sources <chr>, references <chr>,
#> #   curation_effort <dbl>, n_references <chr>, n_resources <int>