Statistics about literature curated ligand-receptor interactions

Statistics about literature curated ligand-receptor interactions

Usage

curated_ligrec_stats(...)

Arguments

...

Passed to curated_ligand_receptor_interactions, determines the set of all curated L-R interactions which will be compared against each of the individual resources.

Value

A data frame with estimated counts of curated ligand-receptor interactions for each L-R resource.

Details

The data frame contains the total number of interactions, the number of interactions which overlap with the set of curated interactions (curated_overlap), the number of interactions with literature references from the given resource (literature) and the number of interactions which are curated by the given resource (curated_self). This latter we defined according to our best knowledge, in many cases it's not possible to distinguish curated interactions). All these numbers are also presented as a percent of the total. Importantly, here we consider interactions curated only if they've been curated in a cell-cell communication context.

See also

Examples

clr <- curated_ligrec_stats()
clr
#> # A tibble: 19 × 8
#>    resource    total curated_overlap literature curated_self curated_overlap_pct
#>    <chr>       <int>           <int>      <int>        <int>               <dbl>
#>  1 CellPhoneDB  2879            2476          0         1905                86.0
#>  2 Cellinker    3765            2341       3765          973                62.2
#>  3 CellTalkDB   3473            3421       3160         3160                98.5
#>  4 CellChatDB   1894            1223        569          569                64.6
#>  5 CellCall     1000             721          0            0                72.1
#>  6 connectome…  2313            2313       2311         2311               100  
#>  7 Guide2Phar…   672             592        299          299                88.1
#>  8 Baccin2019   1676            1383       1381         1381                82.5
#>  9 Kirouac2010   152             122          0            0                80.3
#> 10 Ramilowski…  1946            1904        274          274                97.8
#> 11 scConnect     479             442        291          291                92.3
#> 12 talklr       2484            2365        274          273                95.2
#> 13 ICELLNET     1018            1016       1016         1016                99.8
#> 14 EMBRACE      1707            1620          0            0                94.9
#> 15 LRdb         3310            3135       2052         2052                94.7
#> 16 iTALK        2639            2489          0            0                94.3
#> 17 SignaLink    1796             244       1796          217                13.6
#> 18 HPMR          461             434        416          416                94.1
#> 19 Wojtowicz2…   444             167          0            0                37.6
#> # ℹ 2 more variables: literature_pct <dbl>, curated_self_pct <dbl>