Default values for the package options

These options describe the default settings for OmnipathR so you do not need to pass these parameters at each function call. Currently the only option useful for the public web service at omnipathdb.org is “omnipath.license“. If you are a for-profit user set it to “'commercial'“ to make sure all the data you download from OmniPath is legally allowed for commercial use. Otherwise just leave it as it is: “'academic'“. If you don't use omnipathdb.org but within your organization you deployed your own pypath server and want to share data whith a limited availability to outside users, you may want to use a password. For this you can use the “omnipath.password“ option. Also if you want the R package to work from another pypath server instead of omnipathdb.org, you can change the option “omnipath.url“.

Usage

.omnipath_options_defaults

Format

An object of class list of length 25.

Value

Nothing, this is not a function but a list.