Show evidences for an interaction

Show evidences for an interaction

Usage

evidences(
  partner_a,
  partner_b,
  interactions = NULL,
  directed = FALSE,
  open = TRUE,
  browser = NULL,
  max_pages = 25L
)

Arguments

partner_a

Identifier or name of one interacting partner. The order of the partners matter only if `directed` is `TRUE`. For both partners, vectors of more than one identifiers can be passed.

partner_b

Identifier or name of the other interacting partner.

interactions

An interaction data frame. If not provided, all interactions will be loaded within this function, but that takes noticeable time. If a `list` is provided, it will be used as parameters for omnipath_interactions. This way you can define the organism, datasets or the interaction type.

directed

Logical: does the direction matter? If `TRUE`, only a → b interactions will be shown.

open

Logical: open online articles in a web browser.

browser

Character: override the web browser executable used to open online articles.

max_pages

Numeric: largest number of pages to open. This is to prevent opening hundreds or thousands of pages at once.

Value

Nothing.

Details

If the number of references is larger than `max_pages`, the most recent ones will be opened. URLs are passed to the browser in order of decreasing publication date, though browsers do not seem to respect the order at all. In addition Firefox, if it's not open already, tends to randomly open empty tab for the first or last URL, have no idea what to do about it.

Examples

if (FALSE) { # \dontrun{
evidences('CALM1', 'TRPC1', list(datasets = 'omnipath'))
} # }