Possible values of an extra attribute

Extracts all unique values of an extra attribute occuring in this data frame.

Usage

extra_attr_values(data, key)

Arguments

data

An interaction data frame with extra_attrs column.

key

The name of an extra attribute.

Value

A vector, most likely character, with the unique values of the extra attribute occuring in the data frame.

Details

Note, at the end we unlist the result, which means it works well for attributes which are atomic vectors but gives not so useful result if the attribute values are more complex objects. As the time of writing this, no such complex extra attribute exist in OmniPath.

See also

Examples

op <- omnipath(fields = "extra_attrs")
extra_attr_values(op, SIGNOR_mechanism)
#>  [1] "phosphorylation"                    "binding"                           
#>  [3] "dephosphorylation"                  "Phosphorylation"                   
#>  [5] "ubiquitination"                     "N/A"                               
#>  [7] "Physical Interaction"               "Proteolytic Processing"            
#>  [9] "cleavage"                           "Ubiquitination"                    
#> [11] "Deubiqitination"                    "deubiquitination"                  
#> [13] "relocalization"                     "Dephosphorylation"                 
#> [15] "Other"                              "guanine nucleotide exchange factor"
#> [17] "Transcription Regulation"           "gtpase-activating protein"         
#> [19] "Indirect"                           ""                                  
#> [21] "Sumoylation"                        "sumoylation"                       
#> [23] "palmitoylation"                     "Acetylation"                       
#> [25] "acetylation"                        "polyubiquitination"                
#> [27] "Demethylation"                      "demethylation"                     
#> [29] "mRNA stability"                     "methylation"                       
#> [31] "Methylation"                        "trimethylation"                    
#> [33] "hydroxylation"                      "monoubiquitination"                
#> [35] "Deacetylation"                      "deacetylation"                     
#> [37] "Translational Regulation"           "Protein Degradation"               
#> [39] "Glycosylation"                      "s-nitrosylation"                   
#> [41] "phosphomotif_binding"               "chemical activation"               
#> [43] "Proteolytic Cleavage"               "tyrosination"                      
#> [45] "post transcriptional regulation"    "post translational modification"   
#> [47] "translation regulation"             "carboxylation"                     
#> [49] "neddylation"                        "Carboxylation"                     
#> [51] "desumoylation"                      "glycosylation"                     
#> [53] "ADP-ribosylation"                   "stabilization"                     
#> [55] "catalytic activity"                 "deglycosylation"                   
#> [57] "destabilization"                    "chemical inhibition"               
#> [59] "isomerization"                      "Neddylation"                       
#> [61] "lipidation"                         "chemical modification"             
#> [63] "oxidation"                          "Alkylation"