Retrieve a list of interaction resources available in Omnipath

gets the names of the resources from http://omnipath.org/interactions

get_interaction_resources(dataset = NULL)

get_interaction_databases(...)

Arguments

dataset
a dataset within the interactions query type. Currently available datasets are `omnipath`, `kinaseextra`, `pathwayextra`, `ligrecextra`, `dorothea`, `tf_target`, `tf_mirna`, `mirnatarget` and `lncrna_mrna`

...
Passed to get_interaction_resources.

Value

character vector with the names of the interaction databases

See also

Examples

get_interaction_resources()

#>   [1] "ABS"                       "ACSN"                     
#>   [3] "Adhesome"                  "AlzPathway"               
#>   [5] "ARACNe-GTEx_DoRothEA"      "ARN"                      
#>   [7] "Baccin2019"                "BEL-Large-Corpus"         
#>   [9] "BioGRID"                   "CA1"                      
#>  [11] "CancerCellMap"             "CellPhoneDB"              
#>  [13] "cellsignal.com"            "dbPTM"                    
#>  [15] "DeathDomain"               "DEPOD"                    
#>  [17] "Dinarello2013"             "DIP"                      
#>  [19] "DOMINO"                    "DoRothEA_A"               
#>  [21] "DoRothEA_B"                "DoRothEA_C"               
#>  [23] "DoRothEA_D"                "DoRothEA-reviews_DoRothEA"
#>  [25] "ELM"                       "EMBRACE"                  
#>  [27] "ENCODE_tf-mirna"           "ENCODE-distal"            
#>  [29] "ENCODE-proximal"           "FANTOM4_DoRothEA"         
#>  [31] "Fantom5"                   "GO-lig-rec"               
#>  [33] "Guide2Pharma"              "Guide2Pharma_CellPhoneDB" 
#>  [35] "HOCOMOCO-v11_DoRothEA"     "HPMR"                     
#>  [37] "HPRD"                      "HPRD-phos"                
#>  [39] "HTRIdb"                    "HTRIdb_DoRothEA"          
#>  [41] "I2D_CellPhoneDB"           "ICELLNET"                 
#>  [43] "IMEx_CellPhoneDB"          "InnateDB"                 
#>  [45] "InnateDB_CellPhoneDB"      "InnateDB-All_CellPhoneDB" 
#>  [47] "IntAct_CellPhoneDB"        "IntAct_DoRothEA"          
#>  [49] "iPTMnet"                   "iTALK"                    
#>  [51] "JASPAR-v2018_DoRothEA"     "KEA"                      
#>  [53] "KEGG"                      "KEGG_DoRothEA"            
#>  [55] "Kinexus"                   "Kirouac2010"              
#>  [57] "Li2012"                    "LMPID"                    
#>  [59] "lncrnadb"                  "LncRNADisease"            
#>  [61] "LRdb"                      "Macrophage"               
#>  [63] "MatrixDB"                  "MatrixDB_CellPhoneDB"     
#>  [65] "MIMP"                      "MINT_CellPhoneDB"         
#>  [67] "miR2Disease"               "miRDeathDB"               
#>  [69] "miRecords"                 "miRTarBase"               
#>  [71] "MPPI"                      "NCI-PID"                  
#>  [73] "ncRDeathDB"                "NetPath"                  
#>  [75] "NetworKIN"                 "NFIRegulomeDB_DoRothEA"   
#>  [77] "NRF2ome"                   "ORegAnno"                 
#>  [79] "ORegAnno_DoRothEA"         "PAZAR"                    
#>  [81] "PAZAR_DoRothEA"            "PDZBase"                  
#>  [83] "phosphoELM"                "PhosphoNetworks"          
#>  [85] "PhosphoPoint"              "PhosphoSite"              
#>  [87] "PhosphoSite_noref"         "ProtMapper"               
#>  [89] "Ramilowski2015"            "REACH"                    
#>  [91] "Reactome"                  "RegNetwork_DoRothEA"      
#>  [93] "ReMap_DoRothEA"            "RLIMS-P"                  
#>  [95] "SignaLink3"                "SIGNOR"                   
#>  [97] "Sparser"                   "SPIKE"                    
#>  [99] "STRING"                    "TCRcuration"              
#> [101] "TFactS_DoRothEA"           "TFe_DoRothEA"             
#> [103] "TransmiR"                  "TRED_DoRothEA"            
#> [105] "TRIP"                      "TRRD_DoRothEA"            
#> [107] "TRRUST_DoRothEA"           "UniProt"                  
#> [109] "UniProt_CellPhoneDB"       "Wang"