Retrieve a list of the generic categories in the intercell database of OmniPath

get the names of the generic categories from http://omnipath.org/intercell

get_intercell_generic_categories()

get_intercell_classes(...)

Arguments

...
Passed to get_intercell_generic_categories.

Value

character vector with the different intercell main classes

See also

Examples

get_intercell_generic_categories()

#> Downloaded 1143 records.

#>  [1] "plasma_membrane"                     "transmembrane"                      
#>  [3] "peripheral"                          "transmembrane_predicted"            
#>  [5] "receptor"                            "ligand"                             
#>  [7] "cell_surface_ligand"                 "ecm"                                
#>  [9] "secreted"                            "ion_channel"                        
#> [11] "cell_surface"                        "ligand_regulator"                   
#> [13] "receptor_regulator"                  "plasma_membrane_transmembrane"      
#> [15] "extracellular"                       "secreted_receptor"                  
#> [17] "matrix_adhesion"                     "cell_adhesion"                      
#> [19] "plasma_membrane_peripheral"          "plasma_membrane_regulator"          
#> [21] "secreted_enzyme"                     "cell_surface_enzyme"                
#> [23] "matrix_adhesion_regulator"           "secreted_peptidase"                 
#> [25] "adhesion"                            "transporter"                        
#> [27] "intracellular"                       "cell_surface_peptidase"             
#> [29] "intracellular_intercellular_related" "adherens_junction"                  
#> [31] "ion_channel_regulator"               "extracellular_peptidase"            
#> [33] "tight_junction"                      "secreted_peptidase_inhibitor"       
#> [35] "secreted_enyzme"                     "desmosome"                          
#> [37] "ecm_regulator"                       "gap_junction"                       
#> [39] "sparc_ecm_regulator"