Retrieve the available resources for a given query type

Collects the names of the resources available in OmniPath for a certain query type and optionally for a dataset within that.

get_resources(query_type, datasets = NULL, generic_categories = NULL)

Arguments

query_type
one of the query types `interactions`, `enz_sub`, `complexes`, `annotations` or `intercell`

datasets
currently within the `interactions` query type only, multiple datasets are available: `omnipath`, `kinaseextra`, `pathwayextra`, `ligrecextra`, `dorothea`, `tf_target`, `tf_mirna`, `mirnatarget` and `lncrna_mrna`.

generic_categories
for the `intercell` query type, restrict the search for some generic categories e.g. `ligand` or `receptor`.

Value

a character vector with resource names

Examples

get_resources(query_type = 'interactions')

#>   [1] "ABS"                         "ACSN"                       
#>   [3] "ACSN_SignaLink3"             "Adhesome"                   
#>   [5] "AlzPathway"                  "ARACNe-GTEx_DoRothEA"       
#>   [7] "ARN"                         "Baccin2019"                 
#>   [9] "BEL-Large-Corpus_ProtMapper" "BioGRID"                    
#>  [11] "BioGRID_ICELLNET"            "CA1"                        
#>  [13] "CancerCellMap"               "CellChatDB"                 
#>  [15] "CellChatDB-cofactors"        "CellPhoneDB"                
#>  [17] "CellPhoneDB_ICELLNET"        "CellTalkDB"                 
#>  [19] "connectomeDB2020"            "dbPTM"                      
#>  [21] "DeathDomain"                 "DEPOD"                      
#>  [23] "Dinarello2013_ICELLNET"      "DIP"                        
#>  [25] "DLRP_talklr"                 "DOMINO"                     
#>  [27] "DoRothEA"                    "DoRothEA-reviews_DoRothEA"  
#>  [29] "ELM"                         "EMBRACE"                    
#>  [31] "ENCODE_tf-mirna"             "ENCODE-distal"              
#>  [33] "ENCODE-proximal"             "FANTOM4_DoRothEA"           
#>  [35] "Fantom5_LRdb"                "GO-lig-rec_ICELLNET"        
#>  [37] "Guide2Pharma"                "Guide2Pharma_CellPhoneDB"   
#>  [39] "Guide2Pharma_ICELLNET"       "Guide2Pharma_LRdb"          
#>  [41] "Guide2Pharma_talklr"         "HOCOMOCO_DoRothEA"          
#>  [43] "HPMR"                        "HPMR_ICELLNET"              
#>  [45] "HPMR_LRdb"                   "HPMR_talklr"                
#>  [47] "HPRD"                        "HPRD_KEA"                   
#>  [49] "HPRD_LRdb"                   "HPRD_MIMP"                  
#>  [51] "HPRD_talklr"                 "HPRD-phos"                  
#>  [53] "HTRIdb"                      "HTRIdb_DoRothEA"            
#>  [55] "HuRI"                        "I2D_CellPhoneDB"            
#>  [57] "ICELLNET"                    "IMEx_CellPhoneDB"           
#>  [59] "InnateDB"                    "InnateDB_CellPhoneDB"       
#>  [61] "InnateDB_ICELLNET"           "InnateDB_SignaLink3"        
#>  [63] "InnateDB-All_CellPhoneDB"    "IntAct"                     
#>  [65] "IntAct_CellPhoneDB"          "IntAct_DoRothEA"            
#>  [67] "iPTMnet"                     "iTALK"                      
#>  [69] "JASPAR_DoRothEA"             "KEA"                        
#>  [71] "KEGG-MEDICUS"                "Kinexus_KEA"                
#>  [73] "Kirouac2010"                 "Kirouac2010_ICELLNET"       
#>  [75] "Li2012"                      "Lit-BM-17"                  
#>  [77] "LMPID"                       "lncrnadb"                   
#>  [79] "LncRNADisease"               "LRdb"                       
#>  [81] "Macrophage"                  "Macrophage_ICELLNET"        
#>  [83] "MatrixDB"                    "MatrixDB_CellPhoneDB"       
#>  [85] "MIMP"                        "MINT_CellPhoneDB"           
#>  [87] "miR2Disease"                 "miRDeathDB"                 
#>  [89] "miRecords"                   "miRTarBase"                 
#>  [91] "MPPI"                        "NCI-PID_ProtMapper"         
#>  [93] "ncRDeathDB"                  "NetPath"                    
#>  [95] "NetworKIN_KEA"               "NFIRegulomeDB_DoRothEA"     
#>  [97] "NRF2ome"                     "ORegAnno"                   
#>  [99] "ORegAnno_DoRothEA"           "PAZAR"                      
#> [101] "PAZAR_DoRothEA"              "phosphoELM"                 
#> [103] "phosphoELM_KEA"              "phosphoELM_MIMP"            
#> [105] "PhosphoNetworks"             "PhosphoPoint"               
#> [107] "PhosphoSite"                 "PhosphoSite_KEA"            
#> [109] "PhosphoSite_MIMP"            "PhosphoSite_noref"          
#> [111] "PhosphoSite_ProtMapper"      "ProtMapper"                 
#> [113] "Ramilowski2015"              "Ramilowski2015_Baccin2019"  
#> [115] "Ramilowski2015_ICELLNET"     "REACH_ProtMapper"           
#> [117] "Reactome_ICELLNET"           "Reactome_LRdb"              
#> [119] "Reactome_ProtMapper"         "Reactome_SignaLink3"        
#> [121] "RegNetwork_DoRothEA"         "ReMap_DoRothEA"             
#> [123] "RLIMS-P_ProtMapper"          "SignaLink3"                 
#> [125] "SignaLink3_ICELLNET"         "SIGNOR"                     
#> [127] "SIGNOR_ICELLNET"             "SIGNOR_ProtMapper"          
#> [129] "Sparser_ProtMapper"          "SPIKE"                      
#> [131] "SPIKE_ICELLNET"              "STRING_ICELLNET"            
#> [133] "STRING_talklr"               "talklr"                     
#> [135] "TCRcuration_SignaLink3"      "TFactS_DoRothEA"            
#> [137] "TFe_DoRothEA"                "TransmiR"                   
#> [139] "TRED_DoRothEA"               "TRIP"                       
#> [141] "TRRD_DoRothEA"               "TRRUST_DoRothEA"            
#> [143] "UniProt_CellPhoneDB"         "UniProt_LRdb"               
#> [145] "Wang"                        "Wojtowicz2020"