Gene annotations from Gene Ontology

Gene Ontology is an ontology of gene subcellular localizations, molecular functions and involvement in biological processes. Gene products across many organisms are annotated with the ontology terms. This function downloads the gene-ontology term associations for certain model organisms or all organisms. For a description of the columns see http://geneontology.org/docs/go-annotation-file-gaf-format-2.2/.

Usage

go_annot_download(organism = "human", aspects = c("C", "F", "P"), slim = NULL)

Arguments

organism

Character: either "chicken", "cow", "dog", "human", "pig" or "uniprot_all".

aspects

Character vector with some of the following elements: "C" (cellular component), "F" (molecular function) and "P" (biological process). Gene Ontology is three separate ontologies called as three aspects. By this parameter you can control which aspects to include in the output.

slim

Character: if not NULL, the name of a GOsubset (slim). instead of the full GO annotation, the slim annotation will be returned. See details at go_annot_slim. If TRUE, the "generic" slim will be used.

Value

A tibble (data frame) of annotations as it is provided by the database

Examples

goa_data <- go_annot_download()
goa_data
#> # A tibble: 635,753 × 17
#>    db        db_ob…¹ db_ob…² quali…³ go_id db_ref evide…⁴ with_…⁵ aspect db_ob…⁶
#>    <fct>     <chr>   <chr>   <fct>   <chr> <chr>  <fct>   <chr>   <fct>  <chr>  
#>  1 UniProtKB A0A024… NUDT4B  enables GO:0… GO_RE… IEA     UniPro… F      Diphos…
#>  2 UniProtKB A0A024… NUDT4B  enables GO:0… GO_RE… IEA     UniPro… F      Diphos…
#>  3 UniProtKB A0A024… NUDT4B  locate… GO:0… GO_RE… IDA     NA      C      Diphos…
#>  4 UniProtKB A0A075… IGKV3-7 involv… GO:0… GO_RE… IEA     UniPro… P      Probab…
#>  5 UniProtKB A0A075… IGKV3-7 locate… GO:0… GO_RE… IEA     UniPro… C      Probab…
#>  6 UniProtKB A0A075… IGKV3-7 part_of GO:0… GO_RE… IEA     UniPro… C      Probab…
#>  7 UniProtKB A0A075… IGKV1D… involv… GO:0… GO_RE… IEA     UniPro… P      Probab…
#>  8 UniProtKB A0A075… IGKV1D… locate… GO:0… GO_RE… IEA     UniPro… C      Probab…
#>  9 UniProtKB A0A075… IGKV1D… part_of GO:0… GO_RE… IEA     UniPro… C      Probab…
#> 10 UniProtKB A0A075… IGLV4-… involv… GO:0… GO_RE… IEA     UniPro… P      Immuno…
#> # … with 635,743 more rows, 7 more variables: db_object_synonym <chr>,
#> #   db_object_type <fct>, taxon <fct>, date <date>, assigned_by <fct>,
#> #   annotation_extension <chr>, gene_product_from_id <chr>, and abbreviated
#> #   variable names ¹​db_object_id, ²​db_object_symbol, ³​qualifier,
#> #   ⁴​evidence_code, ⁵​with_or_from, ⁶​db_object_name
# # A tibble: 606,840 x 17
#    db       db_object_id db_object_symbol qualifier go_id   db_ref
#    <fct>    <chr>        <chr>            <fct>     <chr>   <chr>
#  1 UniProt. A0A024RBG1   NUDT4B           NA        GO:000. GO_REF:00.
#  2 UniProt. A0A024RBG1   NUDT4B           NA        GO:000. GO_REF:00.
#  3 UniProt. A0A024RBG1   NUDT4B           NA        GO:004. GO_REF:00.
#  4 UniProt. A0A024RBG1   NUDT4B           NA        GO:005. GO_REF:00.
#  5 UniProt. A0A024RBG1   NUDT4B           NA        GO:005. GO_REF:00.
# # . with 606,830 more rows, and 11 more variables:
# #   evidence_code <fct>, with_or_from <chr>, aspect <fct>,
# #   db_object_name <chr>, db_object_synonym <chr>,
# #   db_object_type <fct>, taxon <fct>, date <date>,
# #   assigned_by <fct>, annotation_extension <chr>,
# #   gene_product_from_id <chr>