Identifier translation table from HMDB

Identifier translation table from HMDB

Usage

hmdb_id_mapping_table(to, from, entity_type = "metabolite")

Arguments

to

Character or symbol: target ID type. See Details for possible values.

from

Character or symbol: source ID type. See Details for possible values.

entity_type

Character: "gene" and "smol" are short symbols for proteins, genes and small molecules respectively. Several other synonyms are also accepted.

Value

A data frame (tibble) with columns `From` and `To`.

Details

The arguments to and from can be provided either as character or as symbol (NSE). Their possible values are either HMDB XML tag names or synonyms listed at id_types.

See also

Examples

hmdb_kegg <- hmdb_id_mapping_table("kegg", "hmdb")
hmdb_kegg
#> # A tibble: 217,920 × 2
#>    From        To    
#>    <chr>       <chr> 
#>  1 HMDB0000001 C01152
#>  2 HMDB0000002 C00986
#>  3 HMDB0000005 C00109
#>  4 HMDB0000008 C05984
#>  5 HMDB0000010 C05299
#>  6 HMDB0000011 C01089
#>  7 HMDB0000012 C00526
#>  8 HMDB0000014 C00881
#>  9 HMDB0000015 C05488
#> 10 HMDB0000016 C03205
#> # ℹ 217,910 more rows