Downloads protein annotations from Human Phenotype Ontology

Human Phenotype Ontology (HPO) provides a standardized vocabulary of phenotypic abnormalities encountered in human disease. Each term in the HPO describes a phenotypic abnormality. HPO currently contains over 13,000 terms and over 156,000 annotations to hereditary diseases. See more at https://hpo.jax.org/app/.

hpo_download()

Value

A tibble (data frame) of annotations as it is provided by the database

Examples

hpo_data <- hpo_download()

#> Error in (function (url, fun, post = NULL, http_param = list(), content_param = list(),     path = NULL, ...) {    op <- options(timeout = 600)    on.exit(options(op))    url_loglevel <- if (getOption("omnipath.print_urls"))         omnipath_console_loglevel()    else logger::INFO    retries <- getOption("omnipath.retry_downloads")    log_level(level = url_loglevel, "Retrieving URL: `%s`", url)    if (!is.null(post) || !is.null(path)) {        reader <- fun        fun <- function(url, post = NULL, ...) {            http_param %<>% ensure_list_2            content_param %<>% ensure_list_2            http_method <- if (is.null(post))                 GET            else POST            http_param$body <- post            if (!is.null(path)) {                http_param %<>% c(list(write_disk(path, overwrite = TRUE)))            }            exec(http_method, url = url, !!!http_param) %>% {                if (is.null(path))                   exec(content, ., !!!content_param) %>% reader(...)                else path            }        }    }    args <- list(...)    args$post <- post    for (attempt in seq(retries)) {        log_trace("Attempt %d/%d: `%s`", attempt, retries, url)        result <- tryCatch(exec(fun, url, !!!args), error = identity)        if (inherits(result, "error")) {            msg <- sprintf("Failed to download `%s` (attempt %d/%d); error: %s",                 url, attempt, retries, conditionMessage(result))            if (attempt == retries) {                log_error(msg)                stop(msg)            }            else {                log_warn(msg)                Sys.sleep(1)            }        }        else {            break        }    }    return(result)})("https://ci.monarchinitiative.org/view/hpo/job/hpo.annotations/lastSuccessfulBuild/artifact/rare-diseases/util/annotation/genes_to_phenotype.txt",     function (file, col_names = TRUE, col_types = NULL, col_select = NULL,         id = NULL, locale = default_locale(), na = c("", "NA"),         quoted_na = TRUE, quote = "\"", comment = "", trim_ws = TRUE,         skip = 0, n_max = Inf, guess_max = min(1000, n_max),         progress = show_progress(), name_repair = "unique", num_threads = readr_threads(),         show_col_types = should_show_types(), skip_empty_rows = TRUE,         lazy = TRUE)     {        tokenizer <- tokenizer_tsv(na = na, quoted_na = quoted_na,             quote = quote, comment = comment, trim_ws = trim_ws,             skip_empty_rows = skip_empty_rows)        if (edition_first()) {            return(read_delimited(file, tokenizer, col_names = col_names,                 col_types = col_types, locale = locale, skip = skip,                 skip_empty_rows = skip_empty_rows, comment = comment,                 n_max = n_max, guess_max = guess_max, progress = progress))        }        vroom::vroom(file, delim = "\t", col_names = col_names,             col_types = col_types, col_select = {                {                  col_select                }            }, id = id, .name_repair = name_repair, locale = locale,             skip = skip, comment = comment, skip_empty_rows = skip_empty_rows,             trim_ws = trim_ws, n_max = n_max, guess_max = guess_max,             progress = progress, show_col_types = show_col_types,             altrep = lazy, num_threads = num_threads)    }, NULL, skip = 1, col_names = c("entrez_gene_id", "entrez_gene_symbol",     "hpo_term_id", "hpo_term_name", "frequency_raw", "frequency_hpo",     "info_gd_source", "gd_source", "disease_id"), col_types = structure(list(        cols = list(), default = structure(list(), class = c("collector_guess",         "collector")), delim = NULL), class = "col_spec"), progress = FALSE): Failed to download `https://ci.monarchinitiative.org/view/hpo/job/hpo.annotations/lastSuccessfulBuild/artifact/rare-diseases/util/annotation/genes_to_phenotype.txt` (attempt 3/3); error: HTTP error 503.

hpo_data

#> Error in eval(expr, envir, enclos): object 'hpo_data' not found

# # A tibble: 231,738 x 9
#    entrez_gene_id entrez_gene_symb. hpo_term_id hpo_term_name
#             <dbl> <chr>             <chr>       <chr>
#  1           8192 CLPP              HP:0000013  Hypoplasia of the ute.
#  2           8192 CLPP              HP:0004322  Short stature
#  3           8192 CLPP              HP:0000786  Primary amenorrhea
#  4           8192 CLPP              HP:0000007  Autosomal recessive i.
#  5           8192 CLPP              HP:0000815  Hypergonadotropic hyp.
# # . with 231,733 more rows, and 5 more variables:
# #   frequency_raw <chr>, frequency_hpo <chr>, info_gd_source <chr>,
# #   gd_source <chr>, disease_id <chr>