Import Omnipath Intercell Data

imports the intercell data stored in Omnipath database from http://omnipathdb.org/intercell. Intercell provides information on the roles in inter-cellular signaling. E.g. if a protein is a ligand, a receptor, an extracellular matrix (ECM) component, etc.

import_Omnipath_intercell(
  from_cache_file = NULL,
  select_categories = get_intercell_categories(),
  select_classes = get_intercell_classes()
)

Arguments

from_cache_file
path to an earlier data file

select_categories
vector containing the categories to be retrieved. All the genes belonging to those categories will be returned. For furter information about the categories see get_intercell_categories

select_classes
vector containing the main classes to be retrieved. All the genes belonging to those classes will be returned. For furter information about the main classes see get_intercell_classes

Value

A dataframe cotaining information about roles in inter-cellular signaling.

See also

Examples

intercell = import_Omnipath_intercell(select_categories=c("ecm"))

#> Downloaded 212296 intercell records

#> removed 211239 intercell records during category/class filtering.