Imports annotations from OmniPath

Imports protein annotations about function, localization, expression, structure and other properties of proteins from OmniPath https://omnipathdb.org/annotations. Note: there might be also a few miRNAs annotated; a vast majority of protein complex annotations are inferred from the annotations of the members: if all members carry the same annotation the complex inherits.

Usage

import_omnipath_annotations(
  proteins = NULL,
  resources = NULL,
  wide = FALSE,
  ...
)

Arguments

proteins

Vector containing the genes or proteins for whom annotations will be retrieved (UniProt IDs or HGNC Gene Symbols or miRBase IDs). It is also possible to donwload annotations for protein complexes. To do so, write 'COMPLEX:' right before the genesymbols of the genes integrating the complex. Check the vignette for examples.

resources

Load the annotations only from these databases. See get_annotation_resources for possible values.

wide

Convert the annotation table to wide format, which corresponds more or less to the original resource. If the data comes from more than one resource a list of wide tables will be returned. See examples at pivot_annotations.

...

Additional arguments.

Value

A data frame containing different gene and complex annotations.

Details

Downloading the full annotations dataset is disabled by default because the size of this data is around 1GB. We recommend to retrieve the annotations for a set of proteins or only from a few resources, depending on your interest. You can always download the full database from https://archive.omnipathdb.org/omnipath_webservice_annotations__recent.tsv using any standard R or readr method.

See also

Examples

annotations <- import_omnipath_annotations(
    proteins = c('TP53', 'LMNA'),
    resources = c('HPA_subcellular')
)