Protein complexes from OmniPath

A comprehensive dataset of protein complexes from the https://omnipathdb.org/complexes endpoint of the OmniPath web service.

Usage

import_omnipath_complexes(...)

Arguments

...

Arguments passed on to omnipath_query

organism

Character or integer: name or NCBI Taxonomy ID of the organism. OmniPath is built of human data, and the web service provides orthology translated interactions and enzyme-substrate relationships for mouse and rat. For other organisms and query types, orthology translation will be called automatically on the downloaded human data before returning the result.

resources

Character vector: name of one or more resources. Restrict the data to these resources. For a complete list of available resources, call the `get_<query_type>_resources` functions for the query type of interst.

datasets

Character vector: name of one or more datasets. In the interactions query type a number of datasets are available. The default is caled "omnipath", and corresponds to the curated causal signaling network published in the 2016 OmniPath paper.

genesymbols

Character or logical: TRUE or FALS or "yes" or "no". Include the `genesymbols` column in the results. OmniPath uses UniProt IDs as the primary identifiers, gene symbols are optional.

fields

Character vector: additional fields to include in the result. For a list of available fields, call `query_info("interactions")`.

default_fields

Logical: if TRUE, the default fields will be included.

silent

Logical: if TRUE, no messages will be printed. By default a summary message is printed upon successful download.

logicals

Character vector: fields to be cast to logical.

format

Character: if "json", JSON will be retrieved and processed into a nested list; any other value will return data frame.

download_args

List: parameters to pass to the download function, which is `readr::read_tsv` by default, and `jsonlite::safe_load`.

references_by_resource

Logical: if TRUE,, in the `references` column the PubMed IDs will be prefixed with the names of the resources they are coming from. If FALSE, the `references` column will be a list of unique PubMed IDs.

add_counts

Logical: if TRUE, the number of references and number of resources for each record will be added to the result.

license

Character: license restrictions. By default, data from resources allowing "academic" use is returned by OmniPath. If you use the data for work in a company, you can provide "commercial" or "for-profit", which will restrict the data to those records which are supported by resources that allow for-profit use.

password

Character: password for the OmniPath web service. You can provide a special password here which enables the use of `license = "ignore"` option, completely bypassing the license filter.

exclude

Character vector: resource or dataset names to be excluded. The data will be filtered after download to remove records of the excluded datasets and resources.

json_param

List: parameters to pass to the `jsonlite::fromJSON` when processing JSON columns embedded in the downloaded data. Such columns are "extra_attrs" and "evidences". These are optional columns which provide a lot of extra details about interactions.

strict_evidences

Logical: reconstruct the "sources" and "references" columns of interaction data frames based on the "evidences" column, strictly filtering them to the queried datasets and resources. Without this, the "sources" and "references" fields for each record might contain information for datasets and resources other than the queried ones, because the downloaded records are a result of a simple filtering of an already integrated data frame.

genesymbol_resource

Character: "uniprot" (default) or "ensembl". The OmniPath web service uses the primary gene symbols as provided by UniProt. By passing "ensembl" here, the UniProt gene symbols will be replaced by the ones used in Ensembl. This translation results in a loss of a few records, and multiplication of another few records due to ambiguous translation.

cache

Logical: use caching, load data from and save to the. The cache directory by default belongs to the user, located in the user's default cache directory, and named "OmnipathR". Find out about it by omnipath_get_cachedir. Can be changed by omnipath_set_cachedir.

Value

A dataframe containing information about complexes

See also

Examples

complexes = import_omnipath_complexes(
    resources = c('CORUM', 'hu.MAP')
)