Imports interactions from the TF-miRNA dataset of OmniPath

Imports the dataset from: https://omnipathdb.org/interactions?datasets=tf_mirna, which contains transcription factor-miRNA gene interactions

Usage

import_tf_mirna_interactions(
  resources = NULL,
  organism = 9606,
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = FALSE,
  ...
)

Arguments

resources

interactions not reported in these databases are removed. See get_interaction_resources for more information.

organism

Interactions are available for human, mouse and rat. Choose among: 9606 human (default), 10116 rat and 10090 Mouse

fields

The user can define here the fields to be added. If used, set the next argument, `default_fields`, to FALSE.

default_fields

whether to include the default fields (columns) for the query type. If FALSE, only the fields defined by the user in the `fields` argument will be added.

references_by_resource

if FALSE, removes the resource name prefixes from the references (PubMed IDs); this way the information which reference comes from which resource will be lost and the PubMed IDs will be unique.

exclude

Character: datasets or resources to exclude.

strict_evidences

Logical: restrict the evidences to the queried datasets and resources. If set to FALSE, the directions and effect signs and references might be based on other datasets and resources.

...

optional additional arguments

Value

A dataframe containing TF-miRNA interactions

See also

Examples

interactions <-
    import_tf_mirna_interactions(
        resources = c('TransmiR')
    )