Imports all TF-target interactions from OmniPath

Imports the dataset from: https://omnipathdb.org/interactions?datasets=tf_target,dorothea, which contains transcription factor-target protein coding gene interactions.

Usage

import_transcriptional_interactions(
  resources = NULL,
  organism = 9606,
  dorothea_levels = c("A", "B"),
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = FALSE,
  ...
)

Arguments

resources

interactions not reported in these databases are removed. See get_interaction_resources for more information.

organism

Interactions are available for human, mouse and rat. Choose among: 9606 human (default), 10116 rat and 10090 Mouse

dorothea_levels

Vector detailing the confidence levels of the interactions to be downloaded. In dorothea, every TF-target interaction has a confidence score ranging from A to E, being A the most reliable interactions. By default we take A and B level interactions (c(A, B)). It is to note that E interactions are not available in OmnipathR.

references_by_resource

if FALSE, removes the resource name prefixes from the references (PubMed IDs); this way the information which reference comes from which resource will be lost and the PubMed IDs will be unique.

exclude

Character: datasets or resources to exclude.

strict_evidences

Logical: restrict the evidences to the queried datasets and resources. If set to FALSE, the directions and effect signs and references might be based on other datasets and resources.

...

Optional additional arguments.

Value

A dataframe containing TF-target interactions.

See also

Examples

grn <-
    import_transcriptional_interactions(
        resources = c('PAZAR', 'ORegAnno', 'DoRothEA')
    )
#> Warning: The following resources are not available: PAZAR. Check the resource names for spelling mistakes.
grn
#> # A tibble: 31,896 × 16
#>    source target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>  <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 P01106 O14746 MYC               TERT                        1              1
#>  2 P84022 P05412 SMAD3             JUN                         1              1
#>  3 Q13485 P05412 SMAD4             JUN                         1              1
#>  4 Q04206 P25445 RELA              FAS                         1              1
#>  5 P08047 P04075 SP1               ALDOA                       1              1
#>  6 P04637 P08069 TP53              IGF1R                       1              0
#>  7 Q05516 P20248 ZBTB16            CCNA2                       1              0
#>  8 Q01196 P08700 RUNX1             IL3                         1              0
#>  9 P42224 P38936 STAT1             CDKN1A                      1              1
#> 10 P40763 P38936 STAT3             CDKN1A                      1              1
#> # ℹ 31,886 more rows
#> # ℹ 10 more variables: is_inhibition <dbl>, consensus_direction <dbl>,
#> #   consensus_stimulation <dbl>, consensus_inhibition <dbl>, sources <chr>,
#> #   references <chr>, curation_effort <dbl>, dorothea_level <chr>,
#> #   n_references <dbl>, n_resources <int>