Looks like a UniProt ID?

This function checks only the format of the IDs, no guarantee that these IDs exist in UniProt.

Usage

is_uniprot(identifiers)

Arguments

identifiers

Character: one or more identifiers (typically a single string, a vector or a data frame column).

Value

Logical: true if all elements in the input (except NAs) looks like valid UniProt IDs. If the input is not a character vector, `FALSE` is returned.

Examples

is_uniprot(all_uniprot_acs())
#> [1] TRUE
# [1] TRUE
is_uniprot("P00533")
#> [1] TRUE
# [1] TRUE
is_uniprot("pizza")
#> [1] FALSE
# [1] FALSE