Find adverse drug-drug interactions in KEGG

See https://www.kegg.jp/kegg/rest/keggapi.html#ddi for details.

Usage

kegg_ddi(...)

Arguments

...

Arguments passed on to kegg_query

operation

Character: one of the KEGG REST API operations.

Value

Data frame (tibble) of four columns with names "drug_a", "drug_b", "interaction" and "mechanism".

Examples

kegg_ddi(c("D00564", "D00100", "D00109"))
#> Error in kegg_query(operation, ...): kegg_query: failed to match arguments against any query template for operation `ddi`. The arguments provided: [[D00564,D00100,D00109]]. Errors for template `dbentries`: In KEGG, <dbentries> must be prefixed with an ID type (database or KEGG organism code), which also can be provided here as an argument name.
#>                     In this case, valid prefixes are [D,ndc,yj]. A sample of the items missing this prefix: [D00564,D00100,D00109].