Information about a KEGG Pathway

Information about a KEGG Pathway

Usage

kegg_info(pathway_id)

Arguments

pathway_id

Character: a KEGG Pathway identifier, e.g. "hsa04710". For a complete list of IDs see kegg_pathway_list.

Value

List with the pathway information.

See also

Examples

kegg_info('map00563')
#> $id
#> [1] "map00563"
#> 
#> $name
#> [1] "Glycosylphosphatidylinositol (GPI)-anchor biosynthesis"
#> 
#> $desc
#> [1] "Cell surface proteins can be attached to the cell membrane via the glycolipid structure called glycosylphosphatidylinositol (GPI) anchor. Hundreds of GPI-anchored proteins have been identified in many eukaryotes ranging from protozoa and fungi to mammals. All protein-linked GPI anchors share a common core structure, characterized by the substructure Man (a1-4) GlcN (a1-6) myo-inositol-1P-lipid. Biosynthesis of GPI anchors proceeds in three stages: (i) preassembly of a GPI precursor in the ER membrane, (ii) attachment of the GPI to the C-terminus of a newly synthesized protein in the lumen of the ER, and (iii) lipid remodeling and/or carbohydrate side-chain modifications in the ER and the Golgi. Defects of GPI anchor biosynthesis gene result in a genetic disorder, paroxysmal nocturnal hemoglobinuria."
#> 
#> $pubmed
#>  [1] "12655644" "11102867" "8253761"  "14517336" "10329735" "14734546"
#>  [7] "12802054" "12807000" "15623507" "15635094" "24364251" "25509331"
#> [13] "19837036" "26563290"
#> 
#> $diseases
#> NULL
#> 
#> $rel_pathways
#> [1] "N-Glycan biosynthesis"          "Glycerophospholipid metabolism"
#> 
#> $module
#> [1] "GPI-anchor biosynthesis, core oligosaccharide [PATH:map00563]"
#>