Download the KEGG Pathways database

Downloads all pathway diagrams in the KEGG Pathways database in KGML format and processes the XML to extract the interactions.

kegg_pathways_download(max_expansion = NULL, simplify = FALSE)

Arguments

max_expansion

Numeric: the maximum number of relations derived from a single relation record. As one entry might represent more than one molecular entities, one relation might yield a large number of relations in the processing. This happens in a combinatorial way, e.g. if the two entries represent 3 and 4 entities, that results 12 relations. If NULL, all relations will be expanded.

simplify

Logical: remove KEGG's internal identifiers and the pathway annotations, keep only unique interactions with direction and effect sign.

Value

A data frame (tibble) of interactions.

See also

Examples

# \donttest{
kegg_pw <- kegg_pathways_download(simplify = TRUE)
#> 
#>   KEGG Pathways: ErbB signaling pathway [--------------------] 2/197, eta: 22s
#> 
#>   KEGG Pathways: Ras signaling pathway  [--------------------] 3/197, eta: 24s
#> 
#>   KEGG Pathways: Rap1 signaling pathway [--------------------] 4/197, eta: 28s
#> 
#>   KEGG Pathways: Wnt signaling pathway  [>-------------------] 5/197, eta: 28s
#> 
#>   KEGG Pathways: Notch signaling pat... [>-------------------] 6/197, eta: 30s
#> 
#>   KEGG Pathways: Hedgehog signaling ... [>-------------------] 7/197, eta: 27s
#> 
#>   KEGG Pathways: TGF-beta signaling ... [>-------------------] 8/197, eta: 27s
#> 
#>   KEGG Pathways: Hippo signaling pat... [>-------------------] 9/197, eta: 27s
#> 
#>   KEGG Pathways: Hippo signaling pat... [>------------------] 10/197, eta: 27s
#> 
#>   KEGG Pathways: VEGF signaling pathway [>------------------] 11/197, eta: 26s
#> 
#>   KEGG Pathways: Apelin signaling pa... [>------------------] 12/197, eta: 25s
#> 
#>   KEGG Pathways: JAK-STAT signaling ... [>------------------] 13/197, eta: 25s
#> 
#>   KEGG Pathways: NF-kappa B signalin... [>------------------] 14/197, eta: 25s
#> 
#>   KEGG Pathways: TNF signaling pathway  [>------------------] 15/197, eta: 26s
#> 
#>   KEGG Pathways: HIF-1 signaling pat... [=>-----------------] 16/197, eta: 26s
#> 
#>   KEGG Pathways: FoxO signaling pathway [=>-----------------] 17/197, eta: 26s
#> 
#>   KEGG Pathways: Calcium signaling p... [=>-----------------] 18/197, eta: 26s
#> 
#>   KEGG Pathways: Phosphatidylinosito... [=>-----------------] 19/197, eta: 26s
#> 
#>   KEGG Pathways: Phospholipase D sig... [=>-----------------] 20/197, eta: 26s
#> 
#>   KEGG Pathways: Sphingolipid signal... [=>-----------------] 21/197, eta: 26s
#> 
#>   KEGG Pathways: cAMP signaling pathway [=>-----------------] 22/197, eta: 26s
#> 
#>   KEGG Pathways: cGMP-PKG signaling ... [=>-----------------] 23/197, eta: 26s
#> 
#>   KEGG Pathways: PI3K-Akt signaling ... [=>-----------------] 24/197, eta: 26s
#> 
#>   KEGG Pathways: AMPK signaling pathway [=>-----------------] 25/197, eta: 26s
#> 
#>   KEGG Pathways: mTOR signaling pathway [==>----------------] 26/197, eta: 26s
#> 
#>   KEGG Pathways: Neuroactive ligand-... [==>----------------] 27/197, eta: 26s
#> 
#>   KEGG Pathways: Cytokine-cytokine r... [==>----------------] 28/197, eta: 26s
#> 
#>   KEGG Pathways: Viral protein inter... [==>----------------] 29/197, eta: 28s
#> 
#>   KEGG Pathways: ECM-receptor intera... [==>----------------] 30/197, eta: 28s
#> 
#>   KEGG Pathways: Cell adhesion molec... [==>----------------] 31/197, eta: 28s
#> 
#>   KEGG Pathways: Endocytosis            [==>----------------] 32/197, eta: 29s
#> 
#>   KEGG Pathways: Phagosome              [==>----------------] 33/197, eta: 28s
#> 
#>   KEGG Pathways: Lysosome               [==>----------------] 34/197, eta: 28s
#> 
#>   KEGG Pathways: Peroxisome             [==>----------------] 35/197, eta: 27s
#> 
#>   KEGG Pathways: Autophagy - animal     [==>----------------] 36/197, eta: 27s
#> 
#>   KEGG Pathways: Mitophagy - animal     [===>---------------] 37/197, eta: 27s
#> 
#>   KEGG Pathways: Cell cycle             [===>---------------] 38/197, eta: 26s
#> 
#>   KEGG Pathways: Apoptosis              [===>---------------] 39/197, eta: 26s
#> 
#>   KEGG Pathways: Apoptosis - multipl... [===>---------------] 40/197, eta: 26s
#> 
#>   KEGG Pathways: Ferroptosis            [===>---------------] 41/197, eta: 26s
#> 
#>   KEGG Pathways: Necroptosis            [===>---------------] 42/197, eta: 25s
#> 
#>   KEGG Pathways: p53 signaling pathway  [===>---------------] 43/197, eta: 25s
#> 
#>   KEGG Pathways: Cellular senescence    [===>---------------] 44/197, eta: 25s
#> 
#>   KEGG Pathways: Focal adhesion         [===>---------------] 45/197, eta: 25s
#> 
#>   KEGG Pathways: Adherens junction      [===>---------------] 46/197, eta: 25s
#> 
#>   KEGG Pathways: Tight junction         [====>--------------] 47/197, eta: 25s
#> 
#>   KEGG Pathways: Gap junction           [====>--------------] 48/197, eta: 25s
#> 
#>   KEGG Pathways: Signaling pathways ... [====>--------------] 49/197, eta: 24s
#> 
#>   KEGG Pathways: Regulation of actin... [====>--------------] 50/197, eta: 24s
#> 
#>   KEGG Pathways: Hematopoietic cell ... [====>--------------] 51/197, eta: 24s
#> 
#>   KEGG Pathways: Complement and coag... [====>--------------] 52/197, eta: 24s
#> 
#>   KEGG Pathways: Platelet activation    [====>--------------] 53/197, eta: 24s
#> 
#>   KEGG Pathways: Neutrophil extracel... [====>--------------] 54/197, eta: 24s
#> 
#>   KEGG Pathways: Toll-like receptor ... [====>--------------] 55/197, eta: 24s
#> 
#>   KEGG Pathways: NOD-like receptor s... [====>--------------] 56/197, eta: 24s
#> 
#>   KEGG Pathways: RIG-I-like receptor... [====>--------------] 57/197, eta: 25s
#> 
#>   KEGG Pathways: Cytosolic DNA-sensi... [=====>-------------] 58/197, eta: 24s
#> 
#>   KEGG Pathways: C-type lectin recep... [=====>-------------] 59/197, eta: 24s
#> 
#>   KEGG Pathways: Natural killer cell... [=====>-------------] 60/197, eta: 24s
#> 
#>   KEGG Pathways: Antigen processing ... [=====>-------------] 61/197, eta: 24s
#> 
#>   KEGG Pathways: T cell receptor sig... [=====>-------------] 62/197, eta: 24s
#> 
#>   KEGG Pathways: Th1 and Th2 cell di... [=====>-------------] 63/197, eta: 24s
#> 
#>   KEGG Pathways: Th17 cell different... [=====>-------------] 64/197, eta: 24s
#> 
#>   KEGG Pathways: IL-17 signaling pat... [=====>-------------] 65/197, eta: 24s
#> 
#>   KEGG Pathways: B cell receptor sig... [=====>-------------] 66/197, eta: 24s
#> 
#>   KEGG Pathways: Fc epsilon RI signa... [=====>-------------] 67/197, eta: 23s
#> 
#>   KEGG Pathways: Fc gamma R-mediated... [======>------------] 68/197, eta: 23s
#> 
#>   KEGG Pathways: Leukocyte transendo... [======>------------] 69/197, eta: 23s
#> 
#>   KEGG Pathways: Intestinal immune n... [======>------------] 70/197, eta: 23s
#> 
#>   KEGG Pathways: Chemokine signaling... [======>------------] 71/197, eta: 22s
#> 
#>   KEGG Pathways: Insulin secretion      [======>------------] 72/197, eta: 22s
#> 
#>   KEGG Pathways: Insulin signaling p... [======>------------] 73/197, eta: 22s
#> 
#>   KEGG Pathways: Glucagon signaling ... [======>------------] 74/197, eta: 22s
#> 
#>   KEGG Pathways: Regulation of lipol... [======>------------] 75/197, eta: 22s
#> 
#>   KEGG Pathways: Adipocytokine signa... [======>------------] 76/197, eta: 21s
#> 
#>   KEGG Pathways: PPAR signaling pathway [======>------------] 77/197, eta: 21s
#> 
#>   KEGG Pathways: GnRH secretion         [=======>-----------] 78/197, eta: 21s
#> 
#>   KEGG Pathways: GnRH signaling pathway [=======>-----------] 79/197, eta: 20s
#> 
#>   KEGG Pathways: Ovarian steroidogen... [=======>-----------] 80/197, eta: 20s
#> 
#>   KEGG Pathways: Estrogen signaling ... [=======>-----------] 81/197, eta: 20s
#> 
#>   KEGG Pathways: Prolactin signaling... [=======>-----------] 82/197, eta: 20s
#> 
#>   KEGG Pathways: Oxytocin signaling ... [=======>-----------] 83/197, eta: 20s
#> 
#>   KEGG Pathways: Relaxin signaling p... [=======>-----------] 84/197, eta: 20s
#> 
#>   KEGG Pathways: Growth hormone synt... [=======>-----------] 85/197, eta: 20s
#> 
#>   KEGG Pathways: Thyroid hormone syn... [=======>-----------] 86/197, eta: 20s
#> 
#>   KEGG Pathways: Thyroid hormone sig... [=======>-----------] 87/197, eta: 19s
#> 
#>   KEGG Pathways: Parathyroid hormone... [=======>-----------] 88/197, eta: 19s
#> 
#>   KEGG Pathways: Melanogenesis          [========>----------] 89/197, eta: 19s
#> 
#>   KEGG Pathways: Renin secretion        [========>----------] 90/197, eta: 19s
#> 
#>   KEGG Pathways: Renin-angiotensin s... [========>----------] 91/197, eta: 19s
#> 
#>   KEGG Pathways: Aldosterone synthes... [========>----------] 92/197, eta: 19s
#> 
#>   KEGG Pathways: Cortisol synthesis ... [========>----------] 93/197, eta: 19s
#> 
#>   KEGG Pathways: Cardiac muscle cont... [========>----------] 94/197, eta: 18s
#> 
#>   KEGG Pathways: Adrenergic signalin... [========>----------] 95/197, eta: 18s
#> 
#>   KEGG Pathways: Vascular smooth mus... [========>----------] 96/197, eta: 18s
#> 
#>   KEGG Pathways: Salivary secretion     [========>----------] 97/197, eta: 18s
#> 
#>   KEGG Pathways: Gastric acid secretion [========>----------] 98/197, eta: 18s
#> 
#>   KEGG Pathways: Pancreatic secretion   [=========>---------] 99/197, eta: 17s
#> 
#>   KEGG Pathways: Bile secretion         [========>---------] 100/197, eta: 17s
#> 
#>   KEGG Pathways: Carbohydrate digest... [========>---------] 101/197, eta: 17s
#> 
#>   KEGG Pathways: Protein digestion a... [========>---------] 102/197, eta: 17s
#> 
#>   KEGG Pathways: Fat digestion and a... [========>---------] 103/197, eta: 17s
#> 
#>   KEGG Pathways: Cholesterol metabolism [=========>--------] 104/197, eta: 17s
#> 
#>   KEGG Pathways: Vitamin digestion a... [=========>--------] 105/197, eta: 16s
#> 
#>   KEGG Pathways: Mineral absorption     [=========>--------] 106/197, eta: 16s
#> 
#>   KEGG Pathways: Vasopressin-regulat... [=========>--------] 107/197, eta: 16s
#> 
#>   KEGG Pathways: Aldosterone-regulat... [=========>--------] 108/197, eta: 16s
#> 
#>   KEGG Pathways: Endocrine and other... [=========>--------] 109/197, eta: 15s
#> 
#>   KEGG Pathways: Proximal tubule bic... [=========>--------] 110/197, eta: 15s
#> 
#>   KEGG Pathways: Collecting duct aci... [=========>--------] 111/197, eta: 15s
#> 
#>   KEGG Pathways: Glutamatergic synapse  [=========>--------] 112/197, eta: 15s
#> 
#>   KEGG Pathways: GABAergic synapse      [=========>--------] 113/197, eta: 15s
#> 
#>   KEGG Pathways: Cholinergic synapse    [=========>--------] 114/197, eta: 14s
#> 
#>   KEGG Pathways: Dopaminergic synapse   [==========>-------] 115/197, eta: 14s
#> 
#>   KEGG Pathways: Serotonergic synapse   [==========>-------] 116/197, eta: 14s
#> 
#>   KEGG Pathways: Long-term potentiation [==========>-------] 117/197, eta: 14s
#> 
#>   KEGG Pathways: Long-term depression   [==========>-------] 118/197, eta: 14s
#> 
#>   KEGG Pathways: Retrograde endocann... [==========>-------] 119/197, eta: 14s
#> 
#>   KEGG Pathways: Synaptic vesicle cycle [==========>-------] 120/197, eta: 13s
#> 
#>   KEGG Pathways: Neurotrophin signal... [==========>-------] 121/197, eta: 13s
#> 
#>   KEGG Pathways: Phototransduction      [==========>-------] 122/197, eta: 13s
#> 
#>   KEGG Pathways: Olfactory transduction [==========>-------] 123/197, eta: 13s
#> 
#>   KEGG Pathways: Taste transduction     [==========>-------] 124/197, eta: 13s
#> 
#>   KEGG Pathways: Inflammatory mediat... [==========>-------] 125/197, eta: 12s
#> 
#>   KEGG Pathways: Axon guidance          [===========>------] 126/197, eta: 12s
#> 
#>   KEGG Pathways: Osteoclast differen... [===========>------] 127/197, eta: 12s
#> 
#>   KEGG Pathways: Longevity regulatin... [===========>------] 128/197, eta: 12s
#> 
#>   KEGG Pathways: Longevity regulatin... [===========>------] 129/197, eta: 12s
#> 
#>   KEGG Pathways: Circadian rhythm       [===========>------] 130/197, eta: 12s
#> 
#>   KEGG Pathways: Circadian entrainment  [===========>------] 131/197, eta: 11s
#> 
#>   KEGG Pathways: Thermogenesis          [===========>------] 132/197, eta: 11s
#> 
#>   KEGG Pathways: Pathways in cancer     [===========>------] 133/197, eta: 11s
#> 
#>   KEGG Pathways: Transcriptional mis... [===========>------] 134/197, eta: 11s
#> 
#>   KEGG Pathways: MicroRNAs in cancer    [===========>------] 135/197, eta: 11s
#> 
#>   KEGG Pathways: Proteoglycans in ca... [===========>------] 136/197, eta: 11s
#> 
#>   KEGG Pathways: Chemical carcinogen... [============>-----] 137/197, eta: 11s
#> 
#>   KEGG Pathways: Chemical carcinogen... [============>-----] 138/197, eta: 10s
#> 
#>   KEGG Pathways: Chemical carcinogen... [============>-----] 139/197, eta: 10s
#> 
#>   KEGG Pathways: Central carbon meta... [============>-----] 140/197, eta: 10s
#> 
#>   KEGG Pathways: Choline metabolism ... [============>-----] 141/197, eta: 10s
#> 
#>   KEGG Pathways: PD-L1 expression an... [============>-----] 142/197, eta: 10s
#> 
#>   KEGG Pathways: Colorectal cancer      [============>-----] 143/197, eta: 10s
#> 
#>   KEGG Pathways: Pancreatic cancer      [============>-----] 144/197, eta:  9s
#> 
#>   KEGG Pathways: Hepatocellular carc... [============>-----] 145/197, eta:  9s
#> 
#>   KEGG Pathways: Gastric cancer         [============>-----] 146/197, eta:  9s
#> 
#>   KEGG Pathways: Glioma                 [============>-----] 147/197, eta:  9s
#> 
#>   KEGG Pathways: Thyroid cancer         [=============>----] 148/197, eta:  9s
#> 
#>   KEGG Pathways: Acute myeloid leukemia [=============>----] 149/197, eta:  8s
#> 
#>   KEGG Pathways: Chronic myeloid leu... [=============>----] 150/197, eta:  8s
#> 
#>   KEGG Pathways: Basal cell carcinoma   [=============>----] 151/197, eta:  8s
#> 
#>   KEGG Pathways: Melanoma               [=============>----] 152/197, eta:  8s
#> 
#>   KEGG Pathways: Renal cell carcinoma   [=============>----] 153/197, eta:  8s
#> 
#>   KEGG Pathways: Bladder cancer         [=============>----] 154/197, eta:  8s
#> 
#>   KEGG Pathways: Prostate cancer        [=============>----] 155/197, eta:  7s
#> 
#>   KEGG Pathways: Endometrial cancer     [=============>----] 156/197, eta:  7s
#> 
#>   KEGG Pathways: Breast cancer          [=============>----] 157/197, eta:  7s
#> 
#>   KEGG Pathways: Small cell lung cancer [=============>----] 158/197, eta:  7s
#> 
#>   KEGG Pathways: Non-small cell lung... [==============>---] 159/197, eta:  7s
#> 
#>   KEGG Pathways: Asthma                 [==============>---] 160/197, eta:  6s
#> 
#>   KEGG Pathways: Systemic lupus eryt... [==============>---] 161/197, eta:  6s
#> 
#>   KEGG Pathways: Rheumatoid arthritis   [==============>---] 162/197, eta:  6s
#> 
#>   KEGG Pathways: Autoimmune thyroid ... [==============>---] 163/197, eta:  6s
#> 
#>   KEGG Pathways: Inflammatory bowel ... [==============>---] 164/197, eta:  6s
#> 
#>   KEGG Pathways: Allograft rejection    [==============>---] 165/197, eta:  5s
#> 
#>   KEGG Pathways: Graft-versus-host d... [==============>---] 166/197, eta:  5s
#> 
#>   KEGG Pathways: Primary immunodefic... [==============>---] 167/197, eta:  5s
#> 
#>   KEGG Pathways: Alzheimer disease      [==============>---] 168/197, eta:  5s
#> 
#>   KEGG Pathways: Parkinson disease      [==============>---] 169/197, eta:  5s
#> 
#>   KEGG Pathways: Amyotrophic lateral... [===============>--] 170/197, eta:  5s
#> 
#>   KEGG Pathways: Huntington disease     [===============>--] 171/197, eta:  4s
#> 
#>   KEGG Pathways: Spinocerebellar ataxia [===============>--] 172/197, eta:  4s
#> 
#>   KEGG Pathways: Prion disease          [===============>--] 173/197, eta:  4s
#> 
#>   KEGG Pathways: Pathways of neurode... [===============>--] 174/197, eta:  4s
#> 
#>   KEGG Pathways: Cocaine addiction      [===============>--] 175/197, eta:  4s
#> 
#>   KEGG Pathways: Amphetamine addiction  [===============>--] 176/197, eta:  4s
#> 
#>   KEGG Pathways: Morphine addiction     [===============>--] 177/197, eta:  4s
#> 
#>   KEGG Pathways: Nicotine addiction     [===============>--] 178/197, eta:  3s
#> 
#>   KEGG Pathways: Alcoholism             [===============>--] 179/197, eta:  3s
#> 
#>   KEGG Pathways: Lipid and atheroscl... [===============>--] 180/197, eta:  3s
#> 
#>   KEGG Pathways: Fluid shear stress ... [================>-] 181/197, eta:  3s
#> 
#>   KEGG Pathways: Hypertrophic cardio... [================>-] 182/197, eta:  3s
#> 
#>   KEGG Pathways: Arrhythmogenic righ... [================>-] 183/197, eta:  2s
#> 
#>   KEGG Pathways: Dilated cardiomyopathy [================>-] 184/197, eta:  2s
#> 
#>   KEGG Pathways: Diabetic cardiomyop... [================>-] 185/197, eta:  2s
#> 
#>   KEGG Pathways: Viral myocarditis      [================>-] 186/197, eta:  2s
#> 
#>   KEGG Pathways: Type II diabetes me... [================>-] 187/197, eta:  2s
#> 
#>   KEGG Pathways: Type I diabetes mel... [================>-] 188/197, eta:  2s
#> 
#>   KEGG Pathways: Maturity onset diab... [================>-] 189/197, eta:  1s
#> 
#>   KEGG Pathways: Non-alcoholic fatty... [================>-] 190/197, eta:  1s
#> 
#>   KEGG Pathways: Insulin resistance     [================>-] 191/197, eta:  1s
#> 
#>   KEGG Pathways: AGE-RAGE signaling ... [=================>] 192/197, eta:  1s
#> 
#>   KEGG Pathways: Cushing syndrome       [=================>] 193/197, eta:  1s
#> 
#>   KEGG Pathways: EGFR tyrosine kinas... [=================>] 194/197, eta:  1s
#> 
#>   KEGG Pathways: Platinum drug resis... [=================>] 195/197, eta:  0s
#> 
#>   KEGG Pathways: Antifolate resistance  [=================>] 196/197, eta:  0s
#> 
#>   KEGG Pathways: Endocrine resistance   [==================] 197/197, eta:  0s
#>                                                                               
#> 
kegg_pw
#> # A tibble: 6,193 x 6
#>    uniprot_source uniprot_target type  effect  genesymbol_sour… genesymbol_targ…
#>    <chr>          <chr>          <chr> <chr>   <chr>            <chr>           
#>  1 Q03113         Q15283         PPrel activa… GNA12            RASA2           
#>  2 Q9Y4G8         P62070         PPrel activa… RAPGEF2          RRAS2           
#>  3 Q13972         P62070         PPrel activa… RASGRF1          RRAS2           
#>  4 O95267         P62070         PPrel activa… RASGRP1          RRAS2           
#>  5 P62834         P15056         PPrel activa… RAP1A            BRAF            
#>  6 P01375         P19438         PPrel activa… TNF              TNFRSF1A        
#>  7 P01583         P14778         PPrel activa… IL1A             IL1R1           
#>  8 P48023         P25445         PPrel activa… FASLG            FAS             
#>  9 P01137         P36897         PPrel activa… TGFB1            TGFBR1          
#> 10 P25445         Q9UER7         PPrel activa… FAS              DAXX            
#> # … with 6,183 more rows
# # A tibble: 6,765 x 6
#    uniprot_source uniprot_target type  effect genesymbol_source
#    <chr>          <chr>          <chr> <chr>  <chr>
#  1 Q03113         Q15283         PPrel activ. GNA12
#  2 Q9Y4G8         P62070         PPrel activ. RAPGEF2
#  3 Q13972         P62070         PPrel activ. RASGRF1
#  4 O95267         P62070         PPrel activ. RASGRP1
#  5 P62834         P15056         PPrel activ. RAP1A
# # . with 6,760 more rows, and 1 more variable: genesymbol_target <chr>
# }