Remove prefix from KEGG foreign database identifiers

Remove prefix from KEGG foreign database identifiers

Usage

kegg_rm_prefix(data, ..., .to_names = TRUE)

Arguments

data

A data frame (tibble) with identifier column(s).

...

Columns where the prefixes should be removed, as a tidyselect selection. If empty, everything() is used to select all columns.

.to_names

Logical: if TRUE, the column names will be updated to reflect the removed prefixes.

Value

A data frame (tibble) with the prefixes removed.

Examples

kegg_rm_prefix(kegg_list("ncbi-geneid", "hsa"))
#>  [1] "map"     "ko"      "ec"      "rn"      "<org>"   "br"      "jp"     
#>  [8] "ko"      "<org>"   "M"       "<org>_M" "T"       "C"       "G"      
#> [15] "R"       "RC"      "H"       "D"       "DG"      "N"      
#>  [1] "map"   "ko"    "ec"    "rn"    "hsa"   "br"    "jp"    "ko"    "hsa"  
#> [10] "M"     "hsa_M" "T"     "C"     "G"     "R"     "RC"    "H"     "D"    
#> [19] "DG"    "N"    
#> Error in (function (url, curlopt = list(), callback = NULL, compr = NULL,     ...) {    log_trace("HTTP request by `omnipath_curl`.")    handle <- exec(omnipath_new_handle, !!!curlopt)    compr %<>% get_compr(url, .)    if (!is.null(compr)) {        path <- url %>% fname_from_url %>% file.path(tempdir(),             .)        log_trace(paste0("`%s` compressed file, downloading to ",             "temporary location: `%s`"), compr, path)        curl_call <- quote(curl_fetch_disk(url, path, handle = handle))    }    else {        curl_call <- quote(curl(url, open = "rb", handle = handle))    }    con <- tryCatch(eval(curl_call), error = identity, warning = identity)    log_curl_stats(handle, url)    if (inherits(con, "condition")) {        msg <- conditionMessage(con)        log_warn("HTTP request failed with error: %s", msg)        stop(msg)    }    if (!is.null(compr)) {        con <- path %>% decompr(compr)    }    if (!is.null(callback)) {        result <- tryCatch({            callback(con, ...)        }, error = function(e) {            stop(e)        }, finally = {            close(con)        })        return(result)    }    return(con)})("https://rest.kegg.jp/list/ncbi-geneid", curlopt = list(),     callback = function (file, col_names = TRUE, col_types = NULL,         col_select = NULL, id = NULL, locale = default_locale(),         na = c("", "NA"), quoted_na = TRUE, quote = "\"", comment = "",         trim_ws = TRUE, skip = 0, n_max = Inf, guess_max = min(1000,             n_max), progress = show_progress(), name_repair = "unique",         num_threads = readr_threads(), show_col_types = should_show_types(),         skip_empty_rows = TRUE, lazy = should_read_lazy())     {        tokenizer <- tokenizer_tsv(na = na, quoted_na = quoted_na,             quote = quote, comment = comment, trim_ws = trim_ws,             skip_empty_rows = skip_empty_rows)        if (edition_first()) {            return(read_delimited(file, tokenizer, col_names = col_names,                 col_types = col_types, locale = locale, skip = skip,                 skip_empty_rows = skip_empty_rows, comment = comment,                 n_max = n_max, guess_max = guess_max, progress = progress,                 show_col_types = show_col_types))        }        vroom::vroom(file, delim = "\t", col_names = col_names,             col_types = col_types, col_select = {                {                  col_select                }            }, id = id, .name_repair = name_repair, skip = skip,             n_max = n_max, na = na, quote = quote, comment = comment,             skip_empty_rows = skip_empty_rows, trim_ws = trim_ws,             escape_double = TRUE, escape_backslash = FALSE, locale = locale,             guess_max = guess_max, show_col_types = show_col_types,             progress = progress, altrep = lazy, num_threads = num_threads)    }, col_names = c("id", "name"), col_types = structure(list(        cols = list(), default = structure(list(), class = c("collector_guess",         "collector")), delim = NULL), class = "col_spec"), show_col_types = FALSE,     progress = FALSE): Failed to open 'https://rest.kegg.jp/list/ncbi-geneid': The requested URL returned error: 400