Load a built in database

Load a built in database

load_db(key, param = list())

Arguments

key
Character: the key of the database to load. For a list of available keys see omnipath_show_db.

param
List: override the defaults or pass further parameters to the database loader function. See the loader functions and their default parameters in omnipath_show_db.

Value

Returns NULL.

Details

This function loads a database which is stored within the package namespace until its expiry. The loaded database is accessible by get_db and the loading process is typically initiated by get_db, not by the users directly.

See also

Examples

load_db('go_slim')
omnipath_show_db()

#> # A tibble: 16 x 10
#>    name   last_used           lifetime package  loader loader_param latest_param
#>    <chr>  <dttm>                 <dbl> <chr>    <chr>  <list>       <list>      
#>  1 Gene … 2021-07-23 15:53:33      300 Omnipat… go_on… <named list… <named list…
#>  2 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  3 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  4 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  5 Gene … 2021-07-23 15:54:19      300 Omnipat… go_on… <named list… <named list…
#>  6 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  7 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  8 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  9 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 10 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 11 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 12 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 13 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 14 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 15 GO an… NA                       300 Omnipat… go_an… <named list… <lgl [1]>   
#> 16 UniPr… NA                       300 Omnipat… unipr… <named list… <lgl [1]>   
#> # … with 3 more variables: loaded <lgl>, db <list>, key <chr>