Builds ligand-receptor network for NicheNet using OmniPath

Retrieves network prior knowledge from OmniPath and provides it in a format suitable for NicheNet. This method never downloads the `ligrecextra` dataset because the ligand-receptor interactions are supposed to come from nichenet_lr_network_omnipath.

nichenet_lr_network_omnipath(quality_filter_param = list(), ...)

Arguments

quality_filter_param

List with arguments for filter_intercell_network. It is recommended to check these parameters and apply some quality filtering. The defaults already ensure certain filtering, but you might want more relaxed or stringent options.

...

Passed to import_intercell_network

Value

A network data frame (tibble) with ligand-receptor interactions suitable for use with NicheNet.

See also

Examples

# use only ligand-receptor interactions (not for example ECM-adhesion):
op_lr_network <- nichenet_lr_network_omnipath(ligand_receptor = TRUE)

# use only CellPhoneDB and Guide to Pharmacology:
op_lr_network <- nichenet_lr_network_omnipath(
    resources = c('CellPhoneDB', 'Guide2Pharma')
)

# only interactions where the receiver is a transporter:
op_lr_network <- nichenet_lr_network_omnipath(
    receiver_param = list(parent = 'transporter')
)