Orthologous gene pairs between two organisms

From the web API of Orthologous Matrix (OMA). Items which could not be translated to `id_type` (but present in the data with their internal OMA IDs) are removed.

Usage

oma_pairwise(
  organism_a = "human",
  organism_b = "mouse",
  id_type = "uniprot",
  mappings = c("1:1", "1:m", "n:1", "n:m"),
  only_ids = TRUE
)

Arguments

organism_a

Name or identifier of an organism.

organism_b

Name or identifier of another organism.

id_type

The gene or protein identifier to use in the table. For a list of supported ID types see `omnipathr.env$id_types$oma`. In addition, "genesymbol" is supported, in this case oma_pairwise_genesymbols will be called automatically.

mappings

Character vector: control ambiguous mappings:

1:1 - unambiguous

1:m - one-to-many

n:1 - many-to-one

n:m - many-to-many

only_ids

Logical: include only the two identifier columns, not the mapping type and the orthology group columns.

Value

A data frame with orthologous gene pairs.

Examples

oma_pairwise("human", "mouse", "uniprot")
#> # A tibble: 13,006 × 2
#>    id_organism_a id_organism_b
#>    <chr>         <chr>        
#>  1 P21452        P30549       
#>  2 Q8TAD7        P0C913       
#>  3 Q8N392        Q8K0Q5       
#>  4 O60285        Q641K5       
#>  5 Q07065        Q8BMK4       
#>  6 P78371        P80314       
#>  7 Q8N4U5        Q8K1H7       
#>  8 Q8WU20        Q8C180       
#>  9 O95619        Q9CR11       
#> 10 Q9NW08        P59470       
#> # ℹ 12,996 more rows
# # A tibble: 21,753 × 4
#    id_organism_a id_organism_b mapping oma_group
#    <chr>         <chr>         <chr>       <dbl>
#  1 Q15326        Q8R5C8        1:1       1129380
#  2 Q9Y2E4        B2RQ71        1:1        681224
#  3 Q92615        Q6A0A2        1:1       1135087
#  4 Q9BZE4        Q99ME9        1:1       1176239
#  5 Q9BXS1        Q8BFZ6        1:m            NA
# # … with 21,743 more rows