Orthologous pairs of gene symbols between two organisms

The Orthologous Matrix (OMA), a resource of orthologous relationships between genes, doesn't provide gene symbols, the identifier preferred in many bioinformatics pipelines. Hence this function wraps oma_pairwise by translating the identifiers used in OMA to gene symbols. Items that can not be translated to `id_type` (but present in the data with their internal OMA IDs) will be removed. This part is done by . Then, in this function we translate the identifiers to gene symbols.

Usage

oma_pairwise_genesymbols(
  organism_a = "human",
  organism_b = "mouse",
  id_type = "uniprot",
  mappings = c("1:1", "1:m", "n:1", "n:m"),
  only_ids = TRUE
)

Arguments

organism_a

Name or identifier of an organism.

organism_b

Name or identifier of another organism.

id_type

The gene or protein identifier to use in the table. For a list of supported ID types see `omnipath.env$id_types$oma`. These are the identifiers that will be translated to gene symbols.

mappings

Character vector: control ambiguous mappings:

1:1 - unambiguous

1:m - one-to-many

n:1 - many-to-one

n:m - many-to-many

only_ids

Logical: include only the two identifier columns, not the mapping type and the orthology group columns.

Value

A data frame with orthologous gene pairs.

Examples

oma_pairwise_genesymbols("human", "mouse")
#> # A tibble: 62,414 × 2
#>    id_organism_a id_organism_b
#>    <chr>         <chr>        
#>  1 ZMYND11       Zmynd11      
#>  2 LARP4B        Larp4b       
#>  3 GTPBP4        Gtpbp4       
#>  4 IDI2          Idi2         
#>  5 IDI2          Idi2         
#>  6 WDR37         Wdr37        
#>  7 ADARB2        Adarb2       
#>  8 PFKP          Pfkp         
#>  9 PITRM1        Pitrm1       
#> 10 AKR1E2        Akr1e2       
#> # ℹ 62,404 more rows