Orthologous pairs between two organisms for ID types not supported by OMA

The Orthologous Matrix (OMA), a resource of orthologous relationships between genes, doesn't provide gene symbols, the identifier preferred in many bioinformatics pipelines. Hence this function wraps oma_pairwise by translating the identifiers used in OMA to gene symbols. Items that can not be translated to `id_type` (but present in the data with their internal OMA IDs) will be removed. Then, in this function we translate the identifiers to the desired ID type.

Usage

oma_pairwise_translated(
  organism_a = "human",
  organism_b = "mouse",
  id_type = "uniprot",
  oma_id_type = "uniprot_entry",
  mappings = c("1:1", "1:m", "n:1", "n:m"),
  only_ids = TRUE
)

Arguments

organism_a

Name or identifier of an organism.

organism_b

Name or identifier of another organism.

id_type

The gene or protein identifier to use in the table. For a list of supported ID types see `omnipath.env$id_types$oma`. These are the identifiers that will be translated to gene symbols.

oma_id_type

Character: the gene or protein identifier to be queried from OMA. These IDs will be translated to `id_type`.

mappings

Character vector: control ambiguous mappings:

1:1 - unambiguous

1:m - one-to-many

n:1 - many-to-one

n:m - many-to-many

only_ids

Logical: include only the two identifier columns, not the mapping type and the orthology group columns.

Value

A data frame with orthologous gene pairs.

Examples

oma_pairwise_translated("human", "mouse")
#> # A tibble: 13,006 × 2
#>    id_organism_a id_organism_b
#>    <chr>         <chr>        
#>  1 P21452        P30549       
#>  2 Q8TAD7        P0C913       
#>  3 Q8N392        Q8K0Q5       
#>  4 O60285        Q641K5       
#>  5 Q07065        Q8BMK4       
#>  6 P78371        P80314       
#>  7 Q8N4U5        Q8K1H7       
#>  8 Q8WU20        Q8C180       
#>  9 O95619        Q9CR11       
#> 10 Q9NW08        P59470       
#> # ℹ 12,996 more rows