Literature curated signaling pathways

Imports interactions from the `omnipath` dataset of Omnipath, a dataset that inherits most of its design and contents from the original OmniPath core from the 2016 publication. This dataset consists of about 40k interactions.

Usage

omnipath(
  resources = NULL,
  organism = 9606L,
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = FALSE,
  ...
)

Arguments

resources

interactions not reported in these databases are removed. See get_interaction_resources for more information.

organism

Interactions are available for human, mouse and rat. Choose among: 9606 human (default), 10116 rat and 10090 Mouse

fields

The user can define here the fields to be added. If used, set the next argument, `default_fields`, to FALSE.

default_fields

whether to include the default fields (columns) for the query type. If FALSE, only the fields defined by the user in the `fields` argument will be added.

references_by_resource

if FALSE, removes the resource name prefixes from the references (PubMed IDs); this way the information which reference comes from which resource will be lost and the PubMed IDs will be unique.

exclude

Character: datasets or resources to exclude.

strict_evidences

Logical: restrict the evidences to the queried datasets and resources. If set to FALSE, the directions and effect signs and references might be based on other datasets and resources.

...

optional additional arguments, passed to import_omnipath_interactions.

Value

A dataframe of literature curated, post-translational signaling interactions.

See also

Examples

pathways <- omnipath()
pathways
#> # A tibble: 81,529 × 15
#>    source target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>  <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 P0DP23 P48995 CALM1             TRPC1                       1              0
#>  2 P0DP25 P48995 CALM3             TRPC1                       1              0
#>  3 P0DP24 P48995 CALM2             TRPC1                       1              0
#>  4 Q03135 P48995 CAV1              TRPC1                       1              1
#>  5 P14416 P48995 DRD2              TRPC1                       1              1
#>  6 Q99750 P48995 MDFI              TRPC1                       1              0
#>  7 Q14571 P48995 ITPR2             TRPC1                       1              1
#>  8 P29966 P48995 MARCKS            TRPC1                       1              0
#>  9 Q13255 P48995 GRM1              TRPC1                       1              1
#> 10 Q13586 P48995 STIM1             TRPC1                       1              1
#> # ℹ 81,519 more rows
#> # ℹ 9 more variables: is_inhibition <dbl>, consensus_direction <dbl>,
#> #   consensus_stimulation <dbl>, consensus_inhibition <dbl>, sources <chr>,
#> #   references <chr>, curation_effort <dbl>, n_references <int>,
#> #   n_resources <int>