Molecular interactions from OmniPath

The functions listed here all download pairwise, causal molecular interactions from the https://omnipathdb.org/interactions endpoint of the OmniPath web service. They are different only in the type of interactions and the kind of resources and data they have been compiled from. A complete list of these functions is available below, these cover the interaction datasets and types currently available in OmniPath:

Interactions from the https://omnipathdb.org/interactions endpoint of the OmniPath web service. By default, it downloads only the "omnipath" dataset, which corresponds to the curated causal interactions described in Turei et al. 2016.

Imports interactions from the `omnipath` dataset of OmniPath, a dataset that inherits most of its design and contents from the original OmniPath core from the 2016 publication. This dataset consists of about 40k interactions.

Imports the dataset from: https://omnipathdb.org/interactions?datasets=pathwayextra, which contains activity flow interactions without literature reference. The activity flow interactions supported by literature references are part of the `omnipath` dataset.

Imports the dataset from: https://omnipathdb.org/interactions?datasets=kinaseextra, which contains enzyme-substrate interactions without literature reference. The enzyme-substrate interactions supported by literature references are part of the `omnipath` dataset.

Imports the dataset from: https://omnipathdb.org/interactions?datasets=ligrecextra, which contains ligand-receptor interactions without literature reference. The ligand-receptor interactions supported by literature references are part of the `omnipath` dataset.

Imports interactions from all post-translational datasets of OmniPath. The datasets are "omnipath", "kinaseextra", "pathwayextra" and "ligrecextra".

Imports the dataset from: https://omnipathdb.org/interactions?datasets=dorothea which contains transcription factor (TF)-target interactions from DoRothEA https://github.com/saezlab/DoRothEA DoRothEA is a comprehensive resource of transcriptional regulation, consisting of 16 original resources, in silico TFBS prediction, gene expression signatures and ChIP-Seq binding site analysis.

Imports the dataset from: https://omnipathdb.org/interactions?datasets=tf_target, which contains transcription factor-target protein coding gene interactions. Note: this is not the only TF-target dataset in OmniPath, `dorothea` is the other one and the `tf_mirna` dataset provides TF-miRNA gene interactions.

Imports the dataset from: https://omnipathdb.org/interactions?datasets=tf_target,dorothea, which contains transcription factor-target protein coding gene interactions.

CollecTRI is a comprehensive resource of transcriptional regulation, published in 2023, consisting of 14 resources and original literature curation.

Imports the dataset from: https://omnipathdb.org/interactions?datasets=mirnatarget, which contains miRNA-mRNA interactions.

Imports the dataset from: https://omnipathdb.org/interactions?datasets=tf_mirna, which contains transcription factor-miRNA gene interactions

Imports the dataset from: https://omnipathdb.org/interactions?datasets=lncrna_mrna, which contains lncRNA-mRNA interactions

Imports the dataset from: https://omnipathdb.org/interactions?datasets=small_molecule, which contains small molecule-protein interactions. Small molecules can be metabolites, intrinsic ligands or drug compounds.

Usage

omnipath_interactions(...)

omnipath(...)

pathwayextra(...)

kinaseextra(...)

ligrecextra(...)

post_translational(...)

dorothea(dorothea_levels = c("A", "B"), ...)

tf_target(...)

transcriptional(dorothea_levels = c("A", "B"), ...)

collectri(...)

mirna_target(...)

tf_mirna(...)

lncrna_mrna(...)

small_molecule(...)

all_interactions(
  dorothea_levels = c("A", "B"),
  types = NULL,
  fields = NULL,
  exclude = NULL,
  ...
)

Arguments

...

Arguments passed on to omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query, omnipath_query

organism

Character or integer: name or NCBI Taxonomy ID of the organism. OmniPath is built of human data, and the web service provides orthology translated interactions and enzyme-substrate relationships for mouse and rat. For other organisms and query types, orthology translation will be called automatically on the downloaded human data before returning the result.

resources

Character vector: name of one or more resources. Restrict the data to these resources. For a complete list of available resources, call the `<query_type>_resources` functions for the query type of interst.

datasets

Character vector: name of one or more datasets. In the interactions query type a number of datasets are available. The default is caled "omnipath", and corresponds to the curated causal signaling network published in the 2016 OmniPath paper.

genesymbols

Character or logical: TRUE or FALS or "yes" or "no". Include the `genesymbols` column in the results. OmniPath uses UniProt IDs as the primary identifiers, gene symbols are optional.

default_fields

Logical: if TRUE, the default fields will be included.

silent

Logical: if TRUE, no messages will be printed. By default a summary message is printed upon successful download.

logicals

Character vector: fields to be cast to logical.

format

Character: if "json", JSON will be retrieved and processed into a nested list; any other value will return data frame.

download_args

List: parameters to pass to the download function, which is `readr::read_tsv` by default, and `jsonlite::safe_load`.

references_by_resource

Logical: if TRUE,, in the `references` column the PubMed IDs will be prefixed with the names of the resources they are coming from. If FALSE, the `references` column will be a list of unique PubMed IDs.

add_counts

Logical: if TRUE, the number of references and number of resources for each record will be added to the result.

license

Character: license restrictions. By default, data from resources allowing "academic" use is returned by OmniPath. If you use the data for work in a company, you can provide "commercial" or "for-profit", which will restrict the data to those records which are supported by resources that allow for-profit use.

password

Character: password for the OmniPath web service. You can provide a special password here which enables the use of `license = "ignore"` option, completely bypassing the license filter.

json_param

List: parameters to pass to the `jsonlite::fromJSON` when processing JSON columns embedded in the downloaded data. Such columns are "extra_attrs" and "evidences". These are optional columns which provide a lot of extra details about interactions.

strict_evidences

Logical: reconstruct the "sources" and "references" columns of interaction data frames based on the "evidences" column, strictly filtering them to the queried datasets and resources. Without this, the "sources" and "references" fields for each record might contain information for datasets and resources other than the queried ones, because the downloaded records are a result of a simple filtering of an already integrated data frame.

genesymbol_resource

Character: "uniprot" (default) or "ensembl". The OmniPath web service uses the primary gene symbols as provided by UniProt. By passing "ensembl" here, the UniProt gene symbols will be replaced by the ones used in Ensembl. This translation results in a loss of a few records, and multiplication of another few records due to ambiguous translation.

cache

Logical: use caching, load data from and save to the. The cache directory by default belongs to the user, located in the user's default cache directory, and named "OmnipathR". Find out about it by getOption("omnipathr.cachedir"). Can be changed by omnipath_set_cachedir.

dorothea_levels

The confidence levels of the dorothea interactions (TF-target) which range from A to D. Set to A and B by default.

types

Character: interaction types, such as "transcriptional", "post_transcriptional", "post_translational", etc.

fields

Character: additional fields (columns) to be included in the result. For a list of available fields, see query_info.

exclude

Character: names of datasets or resource to be excluded from the result. By deafult, the records supported by only these resources or datasets will be removed from the output. If strict_evidences = TRUE, the resource, reference and causality information in the data frame will be reconstructed to remove all information coming from the excluded resources.

Value

A dataframe of molecular interactions.

A dataframe of literature curated, post-translational signaling interactions.

A dataframe containing activity flow interactions between proteins without literature reference

A dataframe containing enzyme-substrate interactions without literature reference

A dataframe containing ligand-receptor interactions including the ones without literature references

A dataframe containing post-translational interactions

A data frame of TF-target interactions from DoRothEA.

A dataframe containing TF-target interactions

A dataframe containing TF-target interactions.

A dataframe of TF-target interactions.

A dataframe containing miRNA-mRNA interactions

A dataframe containing TF-miRNA interactions

A dataframe containing lncRNA-mRNA interactions

A dataframe of small molecule-protein interactions

A dataframe containing all the datasets in the interactions query

Details

Post-translational (protein-protein, PPI) interactions

omnipath: the OmniPath data as defined in the 2016 paper, an arbitrary optimum between coverage and quality. This dataset contains almost entirely causal (stimulatory or inhibitory; i.e. activity flow , according to the SBGN standard), physical interactions between pairs of proteins, curated by experts from the literature.

pathwayextra: activity flow interactions without literature references.

kinaseextra: enzyme-substrate interactions without literature references.

ligrecextra: ligand-receptor interactions without literature references.

post_translational: all post-translational (protein-protein, PPI) interactions; this is the combination of the omnipath, pathwayextra, kinaseextra and ligrecextra datasets.

TF-target (gene regulatory, GRN) interactions

collectri: transcription factor (TF)-target interactions from CollecTRI.

dorothea: transcription factor (TF)-target interactions from DoRothEA

tf_target: transcription factor (TF)-target interactions from other resources

transcriptional: all transcription factor (TF)-target interactions; this is the combination of the collectri, dorothea and tf_target datasets.

Post-transcriptional (miRNA-target) and other RNA related interactions

In these datasets we intend to collect the literature curated resources, hence we don't include some of the most well known large databases if those are based on predictions or high-throughput assays.

mirna_target: miRNA-mRNA interactions

tf_mirna: TF-miRNA interactions

lncrna_mrna: lncRNA-mRNA interactions

Other interaction access functions

small_molecule: interactions between small molecules and proteins. Currently this is a small, experimental dataset that includes drug-target, ligand-receptor, enzyme-metabolite and other interactions. In the future this will be largely expanded and divided into multiple datasets.

all_interactions: all the interaction datasets combined.

See also

Examples

op <- omnipath(resources = c("CA1", "SIGNOR", "SignaLink3"))
op
#> # A tibble: 63,078 × 15
#>    source target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>  <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 Q13976 Q13507 PRKG1             TRPC3                       1              0
#>  2 Q13976 Q9HCX4 PRKG1             TRPC7                       1              1
#>  3 Q13438 Q9HBA0 OS9               TRPV4                       1              1
#>  4 P18031 Q9H1D0 PTPN1             TRPV6                       1              0
#>  5 P63244 Q9BX84 RACK1             TRPM6                       1              0
#>  6 P17612 Q9GZU1 PRKACA            MCOLN1                      1              0
#>  7 Q16539 P49137 MAPK14            MAPKAPK2                    1              1
#>  8 P49137 Q13151 MAPKAPK2          HNRNPA0                     1              1
#>  9 P49137 O95453 MAPKAPK2          PARN                        1              0
#> 10 O60674 P19235 JAK2              EPOR                        1              1
#> # ℹ 63,068 more rows
#> # ℹ 9 more variables: is_inhibition <dbl>, consensus_direction <dbl>,
#> #   consensus_stimulation <dbl>, consensus_inhibition <dbl>, sources <chr>,
#> #   references <chr>, curation_effort <dbl>, n_references <int>,
#> #   n_resources <int>

interactions = omnipath_interactions(
    resources = "SignaLink3",
    organism = 9606
)

pathways <- omnipath()
pathways
#> # A tibble: 81,529 × 15
#>    source target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>  <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 P0DP23 P48995 CALM1             TRPC1                       1              0
#>  2 P0DP25 P48995 CALM3             TRPC1                       1              0
#>  3 P0DP24 P48995 CALM2             TRPC1                       1              0
#>  4 Q03135 P48995 CAV1              TRPC1                       1              1
#>  5 P14416 P48995 DRD2              TRPC1                       1              1
#>  6 Q99750 P48995 MDFI              TRPC1                       1              0
#>  7 Q14571 P48995 ITPR2             TRPC1                       1              1
#>  8 P29966 P48995 MARCKS            TRPC1                       1              0
#>  9 Q13255 P48995 GRM1              TRPC1                       1              1
#> 10 Q13586 P48995 STIM1             TRPC1                       1              1
#> # ℹ 81,519 more rows
#> # ℹ 9 more variables: is_inhibition <dbl>, consensus_direction <dbl>,
#> #   consensus_stimulation <dbl>, consensus_inhibition <dbl>, sources <chr>,
#> #   references <chr>, curation_effort <dbl>, n_references <int>,
#> #   n_resources <int>

interactions <-
    pathwayextra(
        resources = c("BioGRID", "IntAct"),
        organism = 9606
    )

kinase_substrate <-
   kinaseextra(
       resources = c('PhosphoPoint', 'PhosphoSite'),
       organism = 9606
   )

ligand_receptor <- ligrecextra(
    resources = c('HPRD', 'Guide2Pharma'),
    organism = 9606
)

interactions <- post_translational(resources = "BioGRID")

dorothea_grn <- dorothea(
    resources = c('DoRothEA', 'ARACNe-GTEx_DoRothEA'),
    organism = 9606,
    dorothea_levels = c('A', 'B', 'C')
)
dorothea_grn
#> # A tibble: 32,617 × 16
#>    source target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>  <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 P01106 O14746 MYC               TERT                        1              1
#>  2 P84022 P05412 SMAD3             JUN                         1              1
#>  3 Q13485 P05412 SMAD4             JUN                         1              1
#>  4 P08047 P04075 SP1               ALDOA                       1              1
#>  5 P04637 P08069 TP53              IGF1R                       1              0
#>  6 Q05516 P20248 ZBTB16            CCNA2                       1              0
#>  7 Q01196 P08700 RUNX1             IL3                         1              0
#>  8 P42224 P38936 STAT1             CDKN1A                      1              1
#>  9 P40763 P38936 STAT3             CDKN1A                      1              1
#> 10 Q04206 P08183 RELA              ABCB1                       1              1
#> # ℹ 32,607 more rows
#> # ℹ 10 more variables: is_inhibition <dbl>, consensus_direction <dbl>,
#> #   consensus_stimulation <dbl>, consensus_inhibition <dbl>, sources <chr>,
#> #   references <chr>, curation_effort <dbl>, dorothea_level <chr>,
#> #   n_references <dbl>, n_resources <int>

interactions <- tf_target(resources = c("DoRothEA", "SIGNOR"))

grn <- transcriptional(resources = c("PAZAR", "ORegAnno", "DoRothEA"))
#> Warning: The following resources are not available: PAZAR. Check the resource names for spelling mistakes.
grn
#> # A tibble: 31,896 × 16
#>    source target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>  <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 P01106 O14746 MYC               TERT                        1              1
#>  2 P84022 P05412 SMAD3             JUN                         1              1
#>  3 Q13485 P05412 SMAD4             JUN                         1              1
#>  4 Q04206 P25445 RELA              FAS                         1              1
#>  5 P08047 P04075 SP1               ALDOA                       1              1
#>  6 P04637 P08069 TP53              IGF1R                       1              0
#>  7 Q05516 P20248 ZBTB16            CCNA2                       1              0
#>  8 Q01196 P08700 RUNX1             IL3                         1              0
#>  9 P42224 P38936 STAT1             CDKN1A                      1              1
#> 10 P40763 P38936 STAT3             CDKN1A                      1              1
#> # ℹ 31,886 more rows
#> # ℹ 10 more variables: is_inhibition <dbl>, consensus_direction <dbl>,
#> #   consensus_stimulation <dbl>, consensus_inhibition <dbl>, sources <chr>,
#> #   references <chr>, curation_effort <dbl>, dorothea_level <chr>,
#> #   n_references <dbl>, n_resources <int>

collectri_grn <- collectri()
collectri_grn
#> # A tibble: 64,495 × 15
#>    source  target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>   <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 P01106  O14746 MYC               TERT                        1              1
#>  2 P17947  P02818 SPI1              BGLAP                       1              1
#>  3 COMPLE… P05412 FOSL1_JUNB        JUN                         1              1
#>  4 COMPLE… P05412 FOS_JUN           JUN                         1              1
#>  5 COMPLE… P05412 FOS_JUNB          JUN                         1              1
#>  6 COMPLE… P05412 FOSL2_JUND        JUN                         1              1
#>  7 COMPLE… P05412 FOSL2_JUN         JUN                         1              1
#>  8 COMPLE… P05412 JUN               JUN                         1              1
#>  9 COMPLE… P05412 FOSB_JUNB         JUN                         1              1
#> 10 COMPLE… P05412 JUNB              JUN                         1              1
#> # ℹ 64,485 more rows
#> # ℹ 9 more variables: is_inhibition <dbl>, consensus_direction <dbl>,
#> #   consensus_stimulation <dbl>, consensus_inhibition <dbl>, sources <chr>,
#> #   references <chr>, curation_effort <dbl>, n_references <dbl>,
#> #   n_resources <int>

interactions <- mirna_target( resources = c("miRTarBase", "miRecords"))

interactions <- tf_mirna(resources = "TransmiR")

interactions <- lncrna_mrna(resources = c("ncRDeathDB"))

# What are the targets of aspirin?
interactions <- small_molecule(sources = "ASPIRIN")
# The prostaglandin synthases:
interactions
#> # A tibble: 0 × 14
#> # ℹ 14 variables: source <chr>, target <chr>, source_genesymbol <chr>,
#> #   target_genesymbol <chr>, is_directed <chr>, is_stimulation <chr>,
#> #   is_inhibition <chr>, consensus_direction <chr>,
#> #   consensus_stimulation <chr>, consensus_inhibition <chr>, sources <chr>,
#> #   references <chr>, curation_effort <chr>, n_references <lgl>

interactions <- all_interactions(
    resources = c("HPRD", "BioGRID"),
    organism = 9606
)
#> Warning: The following resources are not available: HPRD, BioGRID. Check the resource names for spelling mistakes.
#> Error in omnipath_check_result(result, url): Failed to download data from OmniPath:
#> URL: https://omnipathdb.org/ineractions?resources=HPRD,BioGRID&datasets=collectri,dorothea,kinaseextra,ligrecextra,lncrna_mrna,mirnatarget,omnipath,pathwayextra,small_molecule,tf_mirna,tf_target,tfregulons&dorothea_levels=A,B&fields=type,dorothea_level&license=academic&types=lncrna_post_transcriptional,mirna_transcriptional,post_transcriptional,post_translational,small_molecule_protein,transcriptional
#> 
#> Failed to download data from OmniPath:
#> URL: http://no-tls.omnipathdb.org/ineractions?resources=HPRD,BioGRID&datasets=collectri,dorothea,kinaseextra,ligrecextra,lncrna_mrna,mirnatarget,omnipath,pathwayextra,small_molecule,tf_mirna,tf_target,tfregulons&dorothea_levels=A,B&fields=type,dorothea_level&license=academic&types=lncrna_post_transcriptional,mirna_transcriptional,post_transcriptional,post_translational,small_molecule_protein,transcriptional
#>