OmniPath PPI for the COSMOS PKN

OmniPath PPI for the COSMOS PKN

Usage

omnipath_for_cosmos(
  organism = 9606L,
  resources = NULL,
  datasets = NULL,
  interaction_types = NULL,
  id_types = c("uniprot", "genesymbol"),
  ...
)

Arguments

organism

Character or integer: name or NCBI Taxonomy ID of the organism.

resources

Character: names of one or more resources. Correct spelling is important.

datasets

Character: one or more network datasets in OmniPath.

interaction_types

Character: one or more interaction type

id_types

Character: translate the protein identifiers to these ID types. Each ID type results two extra columns in the output, for the "source" and "target" sides of the interaction, respectively. The default ID type for proteins is Esembl Gene ID, and by default UniProt IDs and Gene Symbols are included. The UniProt IDs returned by the web service are left intact, while the Gene Symbols are queried from Ensembl. These Gene Symbols are different from the ones returned from the web service, and match the Ensembl Gene Symbols used by other components of the COSMOS PKN.

...

Further parameters to import_omnipath_interactions.

Value

Data frame with the columns source, target and sign.

See also

Examples

op_cosmos <- omnipath_for_cosmos()
#> Error in x[[v]] <- if (v %in% xnames && is.list(x[[v]]) && is.list(val[[v]])) modifyList(x[[v]],     val[[v]], keep.null = keep.null) else val[[v]]: subscript out of bounds
op_cosmos
#> Error in eval(expr, envir, enclos): object 'op_cosmos' not found