Save the current package configuration

Save the current package configuration

Save the configuration of a certain package

Usage

omnipath_save_config(path = NULL, title = "default", local = FALSE)

save_config(path = NULL, title = "default", local = FALSE, pkg = "OmnipathR")

Arguments

path

Path to the config file. Directories and the file will be created if don't exist.

title

Save the config under this title. One config file might contain multiple configurations, each identified by a title.

local

Save into a config file in the current directory instead of a user level config file. When loading, the config in the current directory has priority over the user level config.

pkg

Character: name of the package

Value

Returns `NULL`.

Examples

if (FALSE) { # \dontrun{
# after this, all downloads will default to commercial licenses
# i.e. the resources that allow only academic use will be excluded:
options(omnipathr.license = 'commercial')
omnipath_save_config()
} # }