Built in database definitions

Databases are resources which might be costly to load but can be used many times by functions which usually automatically load and retrieve them from the database manager. Each database has a lifetime and will be unloaded automatically upon expiry.

omnipath_show_db()

Value

A data frame with the built in database definitions.

Examples

database_definitions <- omnipath_show_db()
database_definitions

#> # A tibble: 16 x 10
#>    name   last_used           lifetime package  loader loader_param latest_param
#>    <chr>  <dttm>                 <dbl> <chr>    <chr>  <list>       <list>      
#>  1 Gene … 2021-07-23 15:53:33      300 Omnipat… go_on… <named list… <named list…
#>  2 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  3 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  4 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  5 Gene … 2021-07-23 15:54:19      300 Omnipat… go_on… <named list… <named list…
#>  6 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  7 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  8 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#>  9 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 10 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 11 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 12 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 13 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 14 Gene … NA                       300 Omnipat… go_on… <named list… <lgl [1]>   
#> 15 GO an… NA                       300 Omnipat… go_an… <named list… <lgl [1]>   
#> 16 UniPr… NA                       300 Omnipat… unipr… <named list… <lgl [1]>   
#> # … with 3 more variables: loaded <lgl>, db <list>, key <chr>

# # A tibble: 14 x 10
#    name       last_used           lifetime package  loader    loader_p.
#    <chr>      <dttm>                 <dbl> <chr>    <chr>     <list>
#  1 Gene Onto. 2021-04-04 20:19:15      300 Omnipat. go_ontol. <named l.
#  2 Gene Onto. NA                       300 Omnipat. go_ontol. <named l.
#  3 Gene Onto. NA                       300 Omnipat. go_ontol. <named l.
#  4 Gene Onto. NA                       300 Omnipat. go_ontol. <named l.
#  5 Gene Onto. NA                       300 Omnipat. go_ontol. <named l.
# ... (truncated)
# # . with 4 more variables: latest_param <list>, loaded <lgl>, db <list>,
# #   key <chr>