Interactions from PathwayCommons

PathwayCommons (http://www.pathwaycommons.org/) provides molecular interactions from a number of databases, in either BioPAX or SIF (simple interaction format). This function retrieves all interactions in SIF format. The data is limited to the interacting pair and the type of the interaction.

pathwaycommons_download()

Value

A data frame (tibble) with interactions.

Examples

pc_interactions <- pathwaycommons_download()
 pc_interactions

#> # A tibble: 1,884,849 x 3
#>    from  type                        to    
#>    <chr> <chr>                       <chr> 
#>  1 A1BG  controls-expression-of      A2M   
#>  2 A1BG  interacts-with              ABCC6 
#>  3 A1BG  interacts-with              ACE2  
#>  4 A1BG  interacts-with              ADAM10
#>  5 A1BG  interacts-with              ADAM17
#>  6 A1BG  interacts-with              ADAM9 
#>  7 A1BG  interacts-with              AGO1  
#>  8 A1BG  controls-phosphorylation-of AKT1  
#>  9 A1BG  controls-state-change-of    AKT1  
#> 10 A1BG  interacts-with              ANXA7 
#> # … with 1,884,839 more rows

# # A tibble: 1,884,849 x 3
#    from  type                        to
#    <chr> <chr>                       <chr>
#  1 A1BG  controls-expression-of      A2M
#  2 A1BG  interacts-with              ABCC6
#  3 A1BG  interacts-with              ACE2
#  4 A1BG  interacts-with              ADAM10
#  5 A1BG  interacts-with              ADAM17
# # . with 1,884,839 more rows