Downloads TF-target interactions from ReMap

ReMap (http://remap.univ-amu.fr/) is a database of ChIP-Seq experiments. It provides raw and merged peaks and CRMs (cis regulatory motifs) with their associations to regulators (TFs). TF-target relationships can be derived as it is written in Garcia-Alonso et al. 2019: "For ChIP-seq, we downloaded the binding peaks from ReMap and scored the interactions between each TF and each gene according to the distance between the TFBSs and the genes’ transcription start sites. We evaluated different filtering strategies that consisted of selecting only the top-scoring 100, 200, 500, and 1000 target genes for each TF." (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6673718/#s1title). This function returns the top TF-target relationships as used in DoRothEA: https://github.com/saezlab/dorothea/blob/master/inst/scripts/02_chip_seq.R).

Usage

remap_dorothea_download()

Value

Data frame with TF-target relationships.

See also

Examples

remap_interactions <- remap_dorothea_download()
remap_interactions
#> # A tibble: 136,988 × 2
#>    tf    target    
#>    <chr> <chr>     
#>  1 ADNP  ABCC1     
#>  2 ADNP  ABCC6     
#>  3 ADNP  ABHD5     
#>  4 ADNP  ABT1      
#>  5 ADNP  AC002066.1
#>  6 ADNP  AC002127.4
#>  7 ADNP  AC005514.2
#>  8 ADNP  AC006076.1
#>  9 ADNP  AC006552.1
#> 10 ADNP  AC007319.1
#> # ℹ 136,978 more rows
# # A tibble: 136,988 x 2
#    tf    target
#    <chr> <chr>
#  1 ADNP  ABCC1
#  2 ADNP  ABCC6
#  3 ADNP  ABHD5
#  4 ADNP  ABT1
#  5 ADNP  AC002066.1
# # . with 136,978 more rows