Retrieve the available resources for a given query type

Collects the names of the resources available in OmniPath for a certain query type and optionally for a dataset within that.

Usage

resources(query_type, datasets = NULL, generic_categories = NULL)

Arguments

query_type

one of the query types `interactions`, `enz_sub`, `complexes`, `annotations` or `intercell`

datasets

currently within the `interactions` query type only, multiple datasets are available: `omnipath`, `kinaseextra`, `pathwayextra`, `ligrecextra`, `dorothea`, `tf_target`, `tf_mirna`, `mirnatarget` and `lncrna_mrna`.

generic_categories

for the `intercell` query type, restrict the search for some generic categories e.g. `ligand` or `receptor`.

Value

a character vector with resource names

Examples

resources(query_type = "interactions")
#>   [1] "ABS"                         "ACSN"                       
#>   [3] "ACSN_SignaLink3"             "ARACNe-GTEx_DoRothEA"       
#>   [5] "ARN"                         "Adhesome"                   
#>   [7] "AlzPathway"                  "BEL-Large-Corpus_ProtMapper"
#>   [9] "Baccin2019"                  "BioGRID"                    
#>  [11] "CA1"                         "CancerCellMap"              
#>  [13] "CancerDrugsDB"               "CellCall"                   
#>  [15] "CellChatDB"                  "CellChatDB-cofactors"       
#>  [17] "CellPhoneDB"                 "CellPhoneDB_Cellinker"      
#>  [19] "CellTalkDB"                  "Cellinker"                  
#>  [21] "CollecTRI"                   "Cui2007"                    
#>  [23] "CytReg_CollecTRI"            "DEPOD"                      
#>  [25] "DIP"                         "DLRP_Cellinker"             
#>  [27] "DLRP_talklr"                 "DOMINO"                     
#>  [29] "DeathDomain"                 "DoRothEA"                   
#>  [31] "DoRothEA-A_CollecTRI"        "DoRothEA-reviews_DoRothEA"  
#>  [33] "ELM"                         "EMBRACE"                    
#>  [35] "ENCODE-distal"               "ENCODE-proximal"            
#>  [37] "ENCODE_tf-mirna"             "ExTRI_CollecTRI"            
#>  [39] "FANTOM4_DoRothEA"            "Fantom5_LRdb"               
#>  [41] "GEREDB_CollecTRI"            "GOA_CollecTRI"              
#>  [43] "Guide2Pharma"                "Guide2Pharma_Cellinker"     
#>  [45] "Guide2Pharma_LRdb"           "Guide2Pharma_talklr"        
#>  [47] "HOCOMOCO_DoRothEA"           "HPMR"                       
#>  [49] "HPMR_Cellinker"              "HPMR_LRdb"                  
#>  [51] "HPMR_talklr"                 "HPRD"                       
#>  [53] "HPRD-phos"                   "HPRD_KEA"                   
#>  [55] "HPRD_LRdb"                   "HPRD_MIMP"                  
#>  [57] "HPRD_talklr"                 "HTRI_CollecTRI"             
#>  [59] "HTRIdb"                      "HTRIdb_DoRothEA"            
#>  [61] "HuRI"                        "ICELLNET"                   
#>  [63] "InnateDB"                    "InnateDB_SignaLink3"        
#>  [65] "IntAct"                      "IntAct_CollecTRI"           
#>  [67] "IntAct_DoRothEA"             "JASPAR_DoRothEA"            
#>  [69] "KEA"                         "KEGG-MEDICUS"               
#>  [71] "Kinexus_KEA"                 "Kirouac2010"                
#>  [73] "LRdb"                        "Li2012"                     
#>  [75] "Lit-BM-17"                   "LncRNADisease"              
#>  [77] "MIMP"                        "MPPI"                       
#>  [79] "Macrophage"                  "MatrixDB"                   
#>  [81] "NCI-PID_ProtMapper"          "NFIRegulomeDB_DoRothEA"     
#>  [83] "NRF2ome"                     "NTNU.Curated_CollecTRI"     
#>  [85] "NetPath"                     "NetworKIN_KEA"              
#>  [87] "ORegAnno"                    "ORegAnno_DoRothEA"          
#>  [89] "PAZAR_DoRothEA"              "Pavlidis2021_CollecTRI"     
#>  [91] "PhosphoNetworks"             "PhosphoPoint"               
#>  [93] "PhosphoSite"                 "PhosphoSite_KEA"            
#>  [95] "PhosphoSite_MIMP"            "PhosphoSite_ProtMapper"     
#>  [97] "PhosphoSite_noref"           "ProtMapper"                 
#>  [99] "REACH_ProtMapper"            "RLIMS-P_ProtMapper"         
#> [101] "Ramilowski2015"              "Ramilowski2015_Baccin2019"  
#> [103] "ReMap_DoRothEA"              "Reactome_LRdb"              
#> [105] "Reactome_ProtMapper"         "Reactome_SignaLink3"        
#> [107] "RegNetwork_DoRothEA"         "SIGNOR"                     
#> [109] "SIGNOR_CollecTRI"            "SIGNOR_ProtMapper"          
#> [111] "SPIKE"                       "SPIKE_LC"                   
#> [113] "STRING_talklr"               "SignaLink3"                 
#> [115] "Sparser_ProtMapper"          "TCRcuration_SignaLink3"     
#> [117] "TFactS_CollecTRI"            "TFactS_DoRothEA"            
#> [119] "TFe_DoRothEA"                "TRED_DoRothEA"              
#> [121] "TRIP"                        "TRRD_DoRothEA"              
#> [123] "TRRUST_CollecTRI"            "TRRUST_DoRothEA"            
#> [125] "TransmiR"                    "UniProt_LRdb"               
#> [127] "Wang"                        "Wojtowicz2020"              
#> [129] "connectomeDB2020"            "dbPTM"                      
#> [131] "iPTMnet"                     "iTALK"                      
#> [133] "miR2Disease"                 "miRDeathDB"                 
#> [135] "miRTarBase"                  "miRecords"                  
#> [137] "ncRDeathDB"                  "phosphoELM"                 
#> [139] "phosphoELM_KEA"              "phosphoELM_MIMP"            
#> [141] "scConnect"                   "talklr"