Visualize node neighborhood with SigmaJS

This function takes an OmniPath interaction data frame as input and returns a sigmaJS object for the subgraph formed by the neighbors of a node of interest.

Usage

show_network(interactions, node = NULL)

Arguments

interactions

An OmniPath interaction data frame.

node

The node of interest.

Value

A sigmaJS object, check http://sigmajs.john-coene.com/index.html for further details and customization options.

Examples

if (FALSE) { # \dontrun{
# get interactions from omnipath
interactions <- omnipath()
# create and plot the network containing ATM neighbors
viz_sigmajs_neighborhood(interactions_df = interactions, int_node = "ATM")
} # }