Chemical-protein interactions from STITCH

Chemical-protein interactions from STITCH

Usage

stitch_network(
  organism = "human",
  min_score = 700L,
  protein_ids = c("uniprot", "genesymbol"),
  metabolite_ids = c("hmdb", "kegg"),
  cosmos = FALSE
)

Arguments

organism

Character or integer: name or NCBI Taxonomy ID of an organism. STITCH supports many organisms, please refer to their web site at https://stitch.embl.de/.

min_score

Confidence cutoff used for STITCH connections (700 by default).

protein_ids

Character: translate the protein identifiers to these ID types. Each ID type results two extra columns in the output, for the "a" and "b" sides of the interaction, respectively. The default ID type for proteins is Esembl Protein ID, and by default UniProt IDs and Gene Symbols are included.

metabolite_ids

Character: translate the protein identifiers to these ID types. Each ID type results two extra columns in the output, for the "a" and "b" sides of the interaction, respectively. The default ID type for metabolites is PubChem CID, and HMDB IDs and KEGG IDs are included.

cosmos

Logical: use COSMOS format?

Value

A data frame of STITCH chemical-protein and protein-chemical interactions with their effect signs, and optionally with identifiers translated.

See also

Examples

stn <- stitch_network(protein_ids = 'genesymbol', metabolite_ids = 'hmdb')