Unique intercellular interactions

In the intercellular network data frames produced by intercell_network, by default each pair of annotations for an interaction is represented in a separate row. This function drops the annotations and keeps only the distinct interacting pairs.

Usage

unique_intercell_network(network, ...)

Arguments

network

An intercellular network data frame as produced by intercell_network.

...

Additional columns to keep. Note: if these have multiple values for an interacting pair, only the first row will be preserved.

Value

A data frame with interacting pairs and interaction attributes.

See also

Examples

icn <- intercell_network()
icn_unique <- unique_intercell_network(icn)