Separate evidences by direction and effect sign

Separate evidences by direction and effect sign

Usage

unnest_evidences(data, longer = FALSE, .keep = FALSE)

Arguments

data

An interaction data frame with "evidences" column.

longer

Logical: If TRUE, the "evidences" column is split into rows.

.keep

Logical: keep the "evidences" column. When unnesting to longer data frame, the "evidences" column will contain the unnested evidences, while the original column will be retained under the "all_evidences" name (if `.keep = TRUE`).

Value

The data frame with new columns or new rows by direction and sign.

See also

Examples

if (FALSE) { # \dontrun{
op <- omnipath_interactions(fields = "evidences")
op <- unnest_evidences(op)
colnames(op)
} # }