The OmnipathR package

OmnipathR is an R package built to provide easy access to the data stored in the Omnipath webservice:

http://omnipathdb.org/

The webservice implements a very simple REST style API. This package make requests by the HTTP protocol to retreive the data. Hence, fast Internet access is required for a propser use of OmnipathR.

The package also provides some utility functions to filter, analyse and visualize the data.

Examples

# Download post-translational modifications:
ptms = import_omnipath_enzsub(resources=c("PhosphoSite", "SIGNOR"))

#> Downloaded 14899 enzyme-substrate relationships.

# Download protein-protein interactions
interactions = import_omnipath_interactions(resources=c("SignaLink3"))

#> Downloaded 1808 interactions.

# Convert to igraph objects:
ptms_g = ptms_graph(ptms = ptms )
OPI_g = interaction_graph(interactions = interactions )

# Print some interactions:
print_interactions(head(ptms))

#>         enzyme interaction           substrate    modification n_resources
#> 3 SRC (P12931)       ====>  TERT_Y707 (O14746) phosphorylation           8
#> 5 FYN (P06241)       ====>  FYB1_Y595 (O15117) phosphorylation           7
#> 2 LCK (P06239)       ====> SOCS3_Y221 (O14543) phosphorylation           6
#> 4 FYN (P06241)       ====>  FYB1_Y651 (O15117) phosphorylation           6
#> 1 LCK (P06239)       ====> SOCS3_Y204 (O14543) phosphorylation           6
#> 6 FYN (P06241)       ====>  FYB1_Y625 (O15117) phosphorylation           2

# interactions with references:
print_interactions(tail(ptms),writeRefs=TRUE)

#>                enzyme interaction            substrate    modification
#> 14894 PRKACA (P17612)       ====>  PLCL1_T246 (Q15111) phosphorylation
#> 14897   ULK2 (Q8IYT8)       ====> SEC16A_S846 (O15027) phosphorylation
#> 14898   ULK2 (Q8IYT8)       ====> DENND3_S472 (A2RUS2) phosphorylation
#> 14899   ULK2 (Q8IYT8)       ====> DENND3_S490 (A2RUS2) phosphorylation
#> 14895  NUAK1 (O60285)       ====>   MAPT_S356 (P10636) phosphorylation
#> 14896  NUAK1 (O60285)       ====>   MAPT_S262 (P10636) phosphorylation
#>       n_resources n_references   references
#> 14894           2            1 KEA:16854455
#> 14897           2            0             
#> 14898           2            0             
#> 14899           2            0             
#> 14895           1            0             
#> 14896           1            0             

# find interactions between kinase and substrate:
print_interactions(dplyr::filter(ptms,enzyme_genesymbol=="MAP2K1",
   substrate_genesymbol=="MAPK3"))

#>            enzyme interaction           substrate    modification n_resources
#> 1 MAP2K1 (Q02750)       ====> MAPK3_Y204 (P27361) phosphorylation           8
#> 2 MAP2K1 (Q02750)       ====> MAPK3_T202 (P27361) phosphorylation           8
#> 3 MAP2K1 (Q02750)       ====> MAPK3_Y210 (P27361) phosphorylation           2
#> 4 MAP2K1 (Q02750)       ====> MAPK3_T207 (P27361) phosphorylation           2

# find shortest paths on the directed network between proteins
printPath_es(shortest_paths(OPI_g,from = "TYRO3",to = "STAT3",
   output = 'epath')$epath[[1]],OPI_g)

#> Warning: At structural_properties.c:745 :Couldn't reach some vertices

#> Empty path

#> NULL

# find all shortest paths between proteins
printPath_vs(all_shortest_paths(ptms_g,from = "SRC",to = "STAT1")$res,ptms_g)

#> Pathway 1: SRC -> STAT1