All ancestors in the ontology tree

Starting from the selected nodes, recursively walks the ontology tree until it reaches the root. Collects all visited nodes, which are the ancestors (parents) of the starting nodes.

Usage

ancestors(
  terms,
  db_key = "go_basic",
  ids = TRUE,
  relations = c("is_a", "part_of", "occurs_in", "regulates", "positively_regulates",
    "negatively_regulates")
)

Arguments

terms

Character vector of ontology term IDs or names. A mixture of IDs and names can be provided.

db_key

Character: key to identify the ontology database. For the available keys see omnipath_show_db.

ids

Logical: whether to return IDs or term names.

relations

Character vector of ontology relation types. Only these relations will be used.

Value

Character vector of ontology IDs. If the input terms are all root nodes, NULL is returned. The starting nodes won't be included in the result unless some of them are ancestors of other starting nodes.

Details

Note: this function relies on the database manager, the first call might take long because of the database load process. Subsequent calls within a short period should be faster. See get_ontology_db.

Examples

ancestors('GO:0005035', ids = FALSE)
#> [1] "molecular_function"                       
#> [2] "transmembrane signaling receptor activity"
#> [3] "death receptor activity"                  
#> [4] "signaling receptor activity"              
#> [5] "molecular transducer activity"            
# [1] "molecular_function"
# [2] "transmembrane signaling receptor activity"
# [3] "signaling receptor activity"
# [4] "molecular transducer activity"