Genome scale metabolic model by Wang et al. 2021

Process the GEMs from Wang et al., 2021 (https://github.com/SysBioChalmers) into convenient tables.

Usage

chalmers_gem(organism = "Human", orphans = TRUE)

Arguments

organism

Character or integer: an organism (taxon) identifier. Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus), 10116 (Rattus norvegicus), 7955 (Danio rerio), 7227 (Drosophila melanogaster) and 6239 (Caenorhabditis elegans).

orphans

Logical: include orphan reactions (reactions without known enzyme).

Value

List containing the following elements:

reactions: tibble of reaction data;

metabolites: tibble of metabolite data;

reaction_ids: translation table of reaction identifiers;

metabolite_ids: translation table of metabolite identifiers;

S: Stoichiometric matrix (sparse).

References

Wang H, Robinson JL, Kocabas P, Gustafsson J, Anton M, Cholley PE, Huang S, Gobom J, Svensson T, Uhlen M, Zetterberg H, Nielsen J. Genome-scale metabolic network reconstruction of model animals as a platform for translational research. Proc Natl Acad Sci U S A. 2021 Jul 27;118(30):e2102344118. doi: doi:10.1073/pnas.2102344118 .

See also

Examples

gem <- chalmers_gem()
#> Loading required package: Matrix