Metabolite ID translation tables from Chalmers Sysbio

Metabolite ID translation tables from Chalmers Sysbio

Usage

chalmers_gem_id_mapping_table(to, from = "metabolicatlas", organism = "Human")

Arguments

to

Character: type of ID to translate to, either label used internally in this package, or a column name from "metabolites.tsv" distributed by Chalmers Sysbio. NSE is supported.

from

Character: type of ID to translate from, same format as "to".

organism

Character or integer: name or identifier of the organism. Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus), 10116 (Rattus norvegicu), 7955 (Danio rerio), 7227 (Drosophila melanogaster) and 6239 (Caenorhabditis elegans).

Value

Tibble with two columns, "From" and "To", with the corresponding ID types.

Examples

chalmers_gem_id_mapping_table('metabolicatlas', 'hmdb')
#> # A tibble: 712 × 2
#>    From        To      
#>    <chr>       <chr>   
#>  1 HMDB0006525 MAM00002
#>  2 HMDB0001320 MAM00013
#>  3 HMDB0003945 MAM00051
#>  4 HMDB0006529 MAM00057
#>  5 HMDB0006345 MAM00084
#>  6 HMDB0006248 MAM00097
#>  7 HMDB0013025 MAM00114
#>  8 HMDB0006267 MAM00121
#>  9 HMDB0006464 MAM00126
#> 10 HMDB0006516 MAM00131
#> # ℹ 702 more rows