Retrieves the ConsensusPathDB network

Compiles a table of binary interactions from ConsensusPathDB (http://cpdb.molgen.mpg.de/) and translates the UniProtKB ACs to Gene Symbols.

Usage

consensuspathdb_download(complex_max_size = 4, min_score = 0.9)

Arguments

complex_max_size

Numeric: do not expand complexes with a higher number of elements than this. ConsensusPathDB does not contain conventional interactions but lists of participants, which might be members of complexes. Some records include dozens of participants and expanding them to binary interactions result thousands, sometimes hundreds of thousands of interactions from one single record. At the end, this process consumes >10GB of memory and results rather unusable data, hence it is recommended to limit the complex sizes at some low number.

min_score

Numeric: each record in ConsensusPathDB comes with a confidence score, expressing the amount of evidences. The default value, a minimum score of 0.9 retains approx. the top 30 percent of the interactions.

Value

Data frame (tibble) with interactions.

Examples

cpdb_data <- consensuspathdb_download(
    complex_max_size = 1,
    min_score = .99
)
nrow(cpdb_data)
#> [1] 1241556
# [1] 252302
colnames(cpdb_data)
#>  [1] "databases"    "references"   "uniprot_a"    "confidence"   "X5"          
#>  [6] "X6"           "X7"           "X8"           "X9"           "record_id"   
#> [11] "uniprot_b"    "in_complex"   "genesymbol_a" "genesymbol_b"
# [1] "databases"  "references" "uniprot_a"    "confidence"   "record_id"
# [6] "uniprot_b"  "in_complex" "genesymbol_a" "genesymbol_b"
cpdb_data
#> # A tibble: 1,241,556 × 14
#>    databases   references    uniprot_a confidence X5    X6    X7    X8        X9
#>    <chr>       <chr>         <chr>     <chr>      <chr> <chr> <chr> <chr>  <dbl>
#>  1 Reactome,P… 22307056,128… MK03_HUM… P27361     MAPK3 HGNC… 5595  ENSG… NA    
#>  2 Reactome,P… 12834348,826… FA10_HUM… P00742     F10   HGNC… 2159  ENSG… NA    
#>  3 Reactome,H… 9003757       PROC_HUM… P04070     PROC  HGNC… 5624  ENSG… NA    
#>  4 Reactome,I… 25416956,322… DHYS_HUM… P49366     DHPS  HGNC… 1725  ENSG… NA    
#>  5 HPRD,React… 9521684       ARSA_HUM… P15289     ARSA  HGNC… 410   ENSG… NA    
#>  6 Reactome    NA            SUMF2_HU… Q8NBJ7,Q8… SUMF… HGNC… 2587… ENSG…  0.998
#>  7 Reactome    NA            SUMF1_HU… Q8NBJ7,Q8… SUMF… HGNC… 2587… ENSG…  0.998
#>  8 HPRD,React… 9528788,1254… ANPRA_HU… P16066     NPR1  HGNC… 4881  ENSG… NA    
#>  9 PhosphoPOI… 1660465,1671… ANPRA_HU… P16066,P0… NPR1… HGNC… 4881… ENSG…  0.193
#> 10 PhosphoPOI… 1660465,1671… ANF_HUMAN P16066,P0… NPR1… HGNC… 4881… ENSG…  0.193
#> # … with 1,241,546 more rows, and 5 more variables: record_id <int>,
#> #   uniprot_b <chr>, in_complex <lgl>, genesymbol_a <chr>, genesymbol_b <chr>
# # A tibble: 252,302 x 9
#    databases references uniprot_a confidence record_id uniprot_b in_com
#    <chr>     <chr>      <chr>          <dbl>     <int> <chr>     <lgl>
#  1 Reactome  NA         SUMF2_HU.      1             1 SUMF1_HU. TRUE
#  2 Reactome  NA         SUMF1_HU.      1             1 SUMF2_HU. TRUE
#  3 DIP,Reac. 22210847,. STIM1_HU.      0.998         2 TRPC1_HU. TRUE
#  4 DIP,Reac. 22210847,. TRPC1_HU.      0.998         2 STIM1_HU. TRUE
# # . with 252,292 more rows, and 2 more variables: genesymbol_a <chr>,
# #   genesymbol_b <chr