Curated ligand-receptor interactions

The OmniPath intercell database annotates individual proteins and complexes, and we combine these annotations with network interactions on the client side, using import_intercell_network. The architecture of this database is complex, aiming to cover a broad range of knowledge on various levels of details and confidence. We can use the intercell_consensus_filter and filter_intercell_network functions for automated, data driven quality filtering, in order to enrich the cell-cell communication network in higher confidence interactions. However, for many users, a simple combination of the most established, expert curated ligand-receptor resources, provided by this function, fits better their purpose.

Usage

curated_ligand_receptor_interactions(
  curated_resources = c("Guide2Pharma", "HPMR", "ICELLNET", "Kirouac2010",
    "CellTalkDB", "CellChatDB", "connectomeDB2020"),
  cellphonedb = TRUE,
  cellinker = TRUE,
  talklr = TRUE,
  signalink = TRUE
)

Arguments

curated_resources

Character vector of the resource names which are considered to be expert curated. You can include any post-translational network resource here, but if you include non ligand-receptor or non curated resources, the result will not fulfill the original intention of this function.

cellphonedb

Logical: include the curated interactions from CellPhoneDB (not the whole CellPhoneDB but a subset of it).

cellinker

Logical: include the curated interactions from Cellinker (not the whole Cellinker but a subset of it).

talklr

Logical: include the curated interactions from talklr (not the whole talklr but a subset of it).

signalink

Logical: include the ligand-receptor interactions from SignaLink. These are all expert curated.

Value

A data frame similar to interactions (network) data frames, the source and target partners being ligand and receptor, respectively.

Details

Some resources are a mixture of curated and bulk imported interactions, and sometimes it's not trivial to separate these, we take care of these here. This function does not use the intercell database of OmniPath, but retrieves and filters a handful of network resources. The returned data frame has the layout of interactions (network) data frames, and the source and target partners implicitly correspond to ligand and receptor. The data frame shows all resources and references for all interactions, but each interaction is supported by at least one ligand-receptor resource which is supposed to based on expert curation in a ligand-receptor context.

See also

Examples

lr <- curated_ligand_receptor_interactions()
lr
#> # A tibble: 5,140 × 15
#>    source target source_genesymbol target_genesymbol is_directed is_stimulation
#>    <chr>  <chr>  <chr>             <chr>                   <dbl>          <dbl>
#>  1 P01588 P19235 EPO               EPOR                        1              1
#>  2 Q9H2A7 O00574 CXCL16            CXCR6                       1              1
#>  3 P21583 P10721 KITLG             KIT                         1              1
#>  4 Q07325 P49682 CXCL9             CXCR3                       1              1
#>  5 P13501 P51681 CCL5              CCR5                        1              1
#>  6 P80075 P51681 CCL8              CCR5                        1              1
#>  7 P13236 P51681 CCL4              CCR5                        1              1
#>  8 P05231 P40189 IL6               IL6ST                       1              1
#>  9 P02778 P49682 CXCL10            CXCR3                       1              1
#> 10 P01148 P30968 GNRH1             GNRHR                       1              1
#> # … with 5,130 more rows, and 9 more variables: is_inhibition <dbl>,
#> #   consensus_direction <dbl>, consensus_stimulation <dbl>,
#> #   consensus_inhibition <dbl>, sources <chr>, references <chr>,
#> #   curation_effort <dbl>, n_references <chr>, n_resources <int>