TF-target interactions from DoRothEA

Imports the dataset from: https://omnipathdb.org/interactions?datasets=dorothea which contains transcription factor (TF)-target interactions from DoRothEA https://github.com/saezlab/DoRothEA DoRothEA is a comprehensive resource of transcriptional regulation, consisting of 16 original resources, in silico TFBS prediction, gene expression signatures and ChIP-Seq binding site analysis.

Usage

dorothea(
  resources = NULL,
  organism = 9606,
  dorothea_levels = c("A", "B"),
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  ...
)

Arguments

resources

interactions not reported in these databases are removed. See get_interaction_resources for more information.

organism

Interactions are available for human, mouse and rat. Choose among: 9606 human (default), 10116 rat and 10090 Mouse

dorothea_levels

Vector detailing the confidence levels of the interactions to be downloaded. In dorothea, every TF-target interaction has a confidence score ranging from A to E, being A the most reliable interactions. By default we take A and B level interactions (c(A, B)). It is to note that E interactions are not available in OmnipathR.

fields

The user can define here the fields to be added. If used, set the next argument, `default_fields`, to FALSE.

default_fields

whether to include the default fields (columns) for the query type. If FALSE, only the fields defined by the user in the `fields` argument will be added.

references_by_resource

if FALSE, removes the resource name prefixes from the references (PubMed IDs); this way the information which reference comes from which resource will be lost and the PubMed IDs will be unique.

exclude

Character: datasets or resources to exclude.

...

optional additional arguments

Value

A dataframe of TF-target interactions from DoRothEA

See also

Examples

dorothea_grn <- import_dorothea_interactions(
    resources = c('DoRothEA', 'ARACNe-GTEx_DoRothEA'),
    organism = 9606,
    dorothea_levels = c('A', 'B', 'C')
)
#> Warning: 'import_dorothea_interactions' is deprecated.
#> Use 'dorothea' instead.
#> See help("Deprecated")
dorothea_grn
#> # A tibble: 32,617 × 16
#>    source target sourc…¹ targe…² is_di…³ is_st…⁴ is_in…⁵ conse…⁶ conse…⁷ conse…⁸
#>    <chr>  <chr>  <chr>   <chr>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 P01106 O14746 MYC     TERT          1       1       0       1       1       0
#>  2 P84022 P05412 SMAD3   JUN           1       1       0       1       1       0
#>  3 Q13485 P05412 SMAD4   JUN           1       1       0       1       1       0
#>  4 P08047 P04075 SP1     ALDOA         1       1       0       1       1       0
#>  5 P04637 P08069 TP53    IGF1R         1       0       1       1       0       1
#>  6 Q05516 P20248 ZBTB16  CCNA2         1       0       1       1       0       1
#>  7 Q01196 P08700 RUNX1   IL3           1       0       1       1       0       1
#>  8 P42224 P38936 STAT1   CDKN1A        1       1       0       1       1       0
#>  9 P40763 P38936 STAT3   CDKN1A        1       1       0       1       1       0
#> 10 Q04206 P08183 RELA    ABCB1         1       1       0       1       1       0
#> # … with 32,607 more rows, 6 more variables: sources <chr>, references <chr>,
#> #   curation_effort <dbl>, dorothea_level <chr>, n_references <dbl>,
#> #   n_resources <int>, and abbreviated variable names ¹​source_genesymbol,
#> #   ²​target_genesymbol, ³​is_directed, ⁴​is_stimulation, ⁵​is_inhibition,
#> #   ⁶​consensus_direction, ⁷​consensus_stimulation, ⁸​consensus_inhibition