Identifier translation table from Ensembl

Identifier translation table from Ensembl

Usage

ensembl_id_mapping_table(to, from = "uniprot", organism = 9606)

Arguments

to

Character or symbol: target ID type. See Details for possible values.

from

Character or symbol: source ID type. See Details for possible values.

organism

Character or integer: NCBI Taxonomy ID or name of the organism (by default 9606 for human).

Value

A data frame (tibble) with columns `From` and `To`.

Details

The arguments to and from can be provided either as character or as symbol (NSE). Their possible values are either Ensembl attribute names or synonyms listed at translate_ids.

See also

Examples

ensp_up <- ensembl_id_mapping_table("ensp")
ensp_up
#> # A tibble: 123,514 × 2
#>    From   To             
#>    <chr>  <chr>          
#>  1 P03886 ENSP00000354687
#>  2 P03891 ENSP00000355046
#>  3 P00395 ENSP00000354499
#>  4 P00403 ENSP00000354876
#>  5 P03928 ENSP00000355265
#>  6 P00846 ENSP00000354632
#>  7 P00414 ENSP00000354982
#>  8 P03897 ENSP00000355206
#>  9 P03901 ENSP00000354728
#> 10 P03905 ENSP00000354961
#> # ℹ 123,504 more rows
# # A tibble: 119,129 × 2
#    From   To
#    <chr>  <chr>
#  1 P03886 ENSP00000354687
#  2 P03891 ENSP00000355046
#  3 P00395 ENSP00000354499
#  4 P00403 ENSP00000354876
#  5 P03928 ENSP00000355265
# # . with 119,124 more rows