Interactions from the EVEX database

Downloads interactions from EVEX, a versatile text mining resource (http://evexdb.org). Translates the Entrez Gene IDs to Gene Symbols and combines the interactions and references into a single data frame.

evex_download(
    min_confidence = NULL,
    remove_negatives = TRUE,
    top_confidence = NULL
)

Arguments

min_confidence

Numeric: a threshold for confidence scores. EVEX confidence scores span roughly from -3 to 3. By providing a numeric value in this range the lower confidence interactions can be removed. If NULL no filtering performed.

remove_negatives

Logical: remove the records with the "negation" attribute set.

top_confidence

Confidence cutoff as quantile (a number between 0 and 1). If NULL no filtering performed.

Value

Data frame (tibble) with interactions.

Examples

evex_interactions <- evex_download()
evex_interactions
#> # A tibble: 368,297 x 13
#>    general_event_id source_entrezgene_id target_entrezgene_… confidence negation
#>               <dbl> <chr>                <chr>                    <dbl>    <dbl>
#>  1               98 8651                 6774                    -1.45         0
#>  2              100 8431                 6774                    -1.45         0
#>  3              205 6261                 6263                     0.370        0
#>  4              435 1044                 1045                    -1.09         0
#>  5              440 1045                 1044                    -1.09         0
#>  6              507 3596                 952                     -1.67         0
#>  7              559 5592                 5329                    -0.326        0
#>  8              575 85004                2255                    -1.31         0
#>  9              657 2321                 7422                    -1.32         0
#> 10              664 2321                 7422                    -0.544        0
#> # … with 368,287 more rows, and 8 more variables: speculation <dbl>,
#> #   coarse_type <chr>, coarse_polarity <chr>, refined_type <chr>,
#> #   refined_polarity <chr>, source_genesymbol <chr>, target_genesymbol <chr>,
#> #   references <chr>
# # A tibble: 368,297 x 13
#   general_event_id source_entrezge. target_entrezge. confidence negation
#               <dbl> <chr>            <chr>                 <dbl>    <dbl>
# 1               98 8651             6774                 -1.45         0
# 2              100 8431             6774                 -1.45         0
# 3              205 6261             6263                  0.370        0
# 4              435 1044             1045                 -1.09         0
# . with 368,287 more rows, and 8 more variables: speculation <dbl>,
#   coarse_type <chr>, coarse_polarity <chr>, refined_type <chr>,
#   refined_polarity <chr>, source_genesymbol <chr>,
#   target_genesymbol <chr>, references <chr>