Filter intercell annotations

Filters a data frame retrieved by import_omnipath_intercell.

Usage

filter_intercell(
  data,
  categories = NULL,
  resources = NULL,
  parent = NULL,
  scope = NULL,
  aspect = NULL,
  source = NULL,
  transmitter = NULL,
  receiver = NULL,
  secreted = NULL,
  plasma_membrane_peripheral = NULL,
  plasma_membrane_transmembrane = NULL,
  proteins = NULL,
  causality = NULL,
  topology = NULL,
  ...
)

Arguments

data

An intercell annotation data frame as provided by import_omnipath_intercell.

categories

Character: allow only these values in the category column.

resources

Character: allow records only from these resources.

parent

Character: filter for records with these parent categories.

scope

Character: filter for records with these annotation scopes. Possible values are generic and specific.

aspect

Character: filter for records with these annotation aspects. Possible values are functional and locational.

source

Character: filter for records with these annotation sources. Possible values are composite and resource_specific.

transmitter

Logical: if TRUE only transmitters, if FALSE only non-transmitters will be selected, if NULL it has no effect.

receiver

Logical: works the same way as transmitters.

secreted

Logical: works the same way as transmitters.

plasma_membrane_peripheral

Logical: works the same way as transmitters.

plasma_membrane_transmembrane

Logical: works the same way as transmitters.

proteins

Character: filter for annotations of these proteins. Gene symbols or UniProt IDs can be used.

causality

Character: filter for records with these causal roles. Possible values are transmitter and receiver. The filter applied simultaneously to the transmitter and receiver arguments, it's just a different notation for the same thing.

topology

Character: filter for records with these localization topologies. Possible values are secreced, plasma_membrane_peripheral and plasma_membrane_transmembrane; the shorter notations sec, pmp and pmtm can be used. Has the same effect as the logical type arguments, just uses a different notation.

...

Ignored.

Value

The intercell annotation data frame filtered according to the specified conditions.

Examples

ic <- import_omnipath_intercell()
ic <- filter_intercell(
    ic,
    transmitter = TRUE,
    secreted = TRUE,
    scope = "specific"
)
#> Removed 320421 and kept 3151 records of intercell data.