Retrieve a list of interaction resources available in Omnipath

Gets the names of the resources from https://omnipath.org/interactions.

get_interaction_resources(dataset = NULL)

Arguments

dataset

a dataset within the interactions query type. Currently available datasets are `omnipath`, `kinaseextra`, `pathwayextra`, `ligrecextra`, `dorothea`, `tf_target`, `tf_mirna`, `mirnatarget` and `lncrna_mrna`

Value

character vector with the names of the interaction databases

See also

Examples

get_interaction_resources()
#>   [1] "ABS"                         "ACSN"                       
#>   [3] "ACSN_SignaLink3"             "ARACNe-GTEx_DoRothEA"       
#>   [5] "ARN"                         "Adhesome"                   
#>   [7] "AlzPathway"                  "BEL-Large-Corpus_ProtMapper"
#>   [9] "Baccin2019"                  "BioGRID"                    
#>  [11] "BioGRID_ICELLNET"            "CA1"                        
#>  [13] "CancerCellMap"               "CellChatDB"                 
#>  [15] "CellChatDB-cofactors"        "CellPhoneDB"                
#>  [17] "CellPhoneDB_ICELLNET"        "CellTalkDB"                 
#>  [19] "DEPOD"                       "DIP"                        
#>  [21] "DLRP_talklr"                 "DOMINO"                     
#>  [23] "DeathDomain"                 "Dinarello2013_ICELLNET"     
#>  [25] "DoRothEA"                    "DoRothEA-reviews_DoRothEA"  
#>  [27] "ELM"                         "EMBRACE"                    
#>  [29] "ENCODE-distal"               "ENCODE-proximal"            
#>  [31] "ENCODE_tf-mirna"             "FANTOM4_DoRothEA"           
#>  [33] "Fantom5_LRdb"                "GO-lig-rec_ICELLNET"        
#>  [35] "Guide2Pharma"                "Guide2Pharma_CellPhoneDB"   
#>  [37] "Guide2Pharma_ICELLNET"       "Guide2Pharma_LRdb"          
#>  [39] "Guide2Pharma_talklr"         "HOCOMOCO_DoRothEA"          
#>  [41] "HPMR"                        "HPMR_ICELLNET"              
#>  [43] "HPMR_LRdb"                   "HPMR_talklr"                
#>  [45] "HPRD"                        "HPRD-phos"                  
#>  [47] "HPRD_KEA"                    "HPRD_LRdb"                  
#>  [49] "HPRD_MIMP"                   "HPRD_talklr"                
#>  [51] "HTRIdb"                      "HTRIdb_DoRothEA"            
#>  [53] "HuRI"                        "I2D_CellPhoneDB"            
#>  [55] "ICELLNET"                    "IMEx_CellPhoneDB"           
#>  [57] "InnateDB"                    "InnateDB-All_CellPhoneDB"   
#>  [59] "InnateDB_CellPhoneDB"        "InnateDB_ICELLNET"          
#>  [61] "InnateDB_SignaLink3"         "IntAct"                     
#>  [63] "IntAct_CellPhoneDB"          "IntAct_DoRothEA"            
#>  [65] "JASPAR_DoRothEA"             "KEA"                        
#>  [67] "KEGG-MEDICUS"                "Kinexus_KEA"                
#>  [69] "Kirouac2010"                 "Kirouac2010_ICELLNET"       
#>  [71] "LMPID"                       "LRdb"                       
#>  [73] "Li2012"                      "Lit-BM-17"                  
#>  [75] "LncRNADisease"               "MIMP"                       
#>  [77] "MINT_CellPhoneDB"            "MPPI"                       
#>  [79] "Macrophage"                  "Macrophage_ICELLNET"        
#>  [81] "MatrixDB"                    "MatrixDB_CellPhoneDB"       
#>  [83] "NCI-PID_ProtMapper"          "NFIRegulomeDB_DoRothEA"     
#>  [85] "NRF2ome"                     "NetPath"                    
#>  [87] "NetworKIN_KEA"               "ORegAnno"                   
#>  [89] "ORegAnno_DoRothEA"           "PAZAR"                      
#>  [91] "PAZAR_DoRothEA"              "PhosphoNetworks"            
#>  [93] "PhosphoPoint"                "PhosphoSite"                
#>  [95] "PhosphoSite_KEA"             "PhosphoSite_MIMP"           
#>  [97] "PhosphoSite_ProtMapper"      "PhosphoSite_noref"          
#>  [99] "ProtMapper"                  "REACH_ProtMapper"           
#> [101] "RLIMS-P_ProtMapper"          "Ramilowski2015"             
#> [103] "Ramilowski2015_Baccin2019"   "Ramilowski2015_ICELLNET"    
#> [105] "ReMap_DoRothEA"              "Reactome_ICELLNET"          
#> [107] "Reactome_LRdb"               "Reactome_ProtMapper"        
#> [109] "Reactome_SignaLink3"         "RegNetwork_DoRothEA"        
#> [111] "SIGNOR"                      "SIGNOR_ICELLNET"            
#> [113] "SIGNOR_ProtMapper"           "SPIKE"                      
#> [115] "SPIKE_ICELLNET"              "STRING_ICELLNET"            
#> [117] "STRING_talklr"               "SignaLink3"                 
#> [119] "SignaLink3_ICELLNET"         "Sparser_ProtMapper"         
#> [121] "TCRcuration_SignaLink3"      "TFactS_DoRothEA"            
#> [123] "TFe_DoRothEA"                "TRED_DoRothEA"              
#> [125] "TRIP"                        "TRRD_DoRothEA"              
#> [127] "TRRUST_DoRothEA"             "TransmiR"                   
#> [129] "UniProt_CellPhoneDB"         "UniProt_LRdb"               
#> [131] "Wang"                        "Wojtowicz2020"              
#> [133] "connectomeDB2020"            "dbPTM"                      
#> [135] "iPTMnet"                     "iTALK"                      
#> [137] "lncrnadb"                    "miR2Disease"                
#> [139] "miRDeathDB"                  "miRTarBase"                 
#> [141] "miRecords"                   "ncRDeathDB"                 
#> [143] "phosphoELM"                  "phosphoELM_KEA"             
#> [145] "phosphoELM_MIMP"             "talklr"