Retrieve a list of interaction resources available in Omnipath

Gets the names of the resources from http://omnipath.org/interactions.

get_interaction_resources(dataset = NULL)

get_interaction_databases(...)

Arguments

dataset
a dataset within the interactions query type. Currently available datasets are `omnipath`, `kinaseextra`, `pathwayextra`, `ligrecextra`, `dorothea`, `tf_target`, `tf_mirna`, `mirnatarget` and `lncrna_mrna`

...
Passed to get_interaction_resources.

Value

character vector with the names of the interaction databases

See also

Examples

get_interaction_resources()

#>   [1] "ABS"                         "ACSN"                       
#>   [3] "ACSN_SignaLink3"             "Adhesome"                   
#>   [5] "AlzPathway"                  "ARACNe-GTEx_DoRothEA"       
#>   [7] "ARN"                         "Baccin2019"                 
#>   [9] "BEL-Large-Corpus_ProtMapper" "BioGRID"                    
#>  [11] "BioGRID_ICELLNET"            "CA1"                        
#>  [13] "CancerCellMap"               "CellPhoneDB"                
#>  [15] "CellPhoneDB_ICELLNET"        "dbPTM"                      
#>  [17] "DeathDomain"                 "DEPOD"                      
#>  [19] "Dinarello2013_ICELLNET"      "DIP"                        
#>  [21] "DOMINO"                      "DoRothEA"                   
#>  [23] "DoRothEA-reviews_DoRothEA"   "ELM"                        
#>  [25] "EMBRACE"                     "ENCODE_tf-mirna"            
#>  [27] "ENCODE-distal"               "ENCODE-proximal"            
#>  [29] "FANTOM4_DoRothEA"            "Fantom5_LRdb"               
#>  [31] "GO-lig-rec_ICELLNET"         "Guide2Pharma"               
#>  [33] "Guide2Pharma_CellPhoneDB"    "Guide2Pharma_ICELLNET"      
#>  [35] "Guide2Pharma_LRdb"           "HOCOMOCO_DoRothEA"          
#>  [37] "HPMR"                        "HPMR_ICELLNET"              
#>  [39] "HPMR_LRdb"                   "HPRD"                       
#>  [41] "HPRD_KEA"                    "HPRD_LRdb"                  
#>  [43] "HPRD_MIMP"                   "HPRD-phos"                  
#>  [45] "HTRIdb"                      "HTRIdb_DoRothEA"            
#>  [47] "HuRI"                        "I2D_CellPhoneDB"            
#>  [49] "ICELLNET"                    "IMEx_CellPhoneDB"           
#>  [51] "InnateDB"                    "InnateDB_CellPhoneDB"       
#>  [53] "InnateDB_ICELLNET"           "InnateDB_SignaLink3"        
#>  [55] "InnateDB-All_CellPhoneDB"    "IntAct"                     
#>  [57] "IntAct_CellPhoneDB"          "IntAct_DoRothEA"            
#>  [59] "iPTMnet"                     "iTALK"                      
#>  [61] "JASPAR_DoRothEA"             "KEA"                        
#>  [63] "KEGG-MEDICUS"                "Kinexus_KEA"                
#>  [65] "Kirouac2010"                 "Kirouac2010_ICELLNET"       
#>  [67] "Li2012"                      "Lit-BM-17"                  
#>  [69] "LMPID"                       "lncrnadb"                   
#>  [71] "LncRNADisease"               "LRdb"                       
#>  [73] "Macrophage"                  "Macrophage_ICELLNET"        
#>  [75] "MatrixDB"                    "MatrixDB_CellPhoneDB"       
#>  [77] "MIMP"                        "MINT_CellPhoneDB"           
#>  [79] "miR2Disease"                 "miRDeathDB"                 
#>  [81] "miRecords"                   "miRTarBase"                 
#>  [83] "MPPI"                        "NCI-PID_ProtMapper"         
#>  [85] "ncRDeathDB"                  "NetPath"                    
#>  [87] "NetworKIN_KEA"               "NFIRegulomeDB_DoRothEA"     
#>  [89] "NRF2ome"                     "ORegAnno"                   
#>  [91] "ORegAnno_DoRothEA"           "PAZAR"                      
#>  [93] "PAZAR_DoRothEA"              "phosphoELM"                 
#>  [95] "phosphoELM_KEA"              "phosphoELM_MIMP"            
#>  [97] "PhosphoNetworks"             "PhosphoPoint"               
#>  [99] "PhosphoSite"                 "PhosphoSite_KEA"            
#> [101] "PhosphoSite_MIMP"            "PhosphoSite_noref"          
#> [103] "PhosphoSite_ProtMapper"      "ProtMapper"                 
#> [105] "Ramilowski2015"              "Ramilowski2015_Baccin2019"  
#> [107] "Ramilowski2015_ICELLNET"     "REACH_ProtMapper"           
#> [109] "Reactome_ICELLNET"           "Reactome_LRdb"              
#> [111] "Reactome_ProtMapper"         "Reactome_SignaLink3"        
#> [113] "RegNetwork_DoRothEA"         "ReMap_DoRothEA"             
#> [115] "RLIMS-P_ProtMapper"          "SignaLink3"                 
#> [117] "SignaLink3_ICELLNET"         "SIGNOR"                     
#> [119] "SIGNOR_ICELLNET"             "SIGNOR_ProtMapper"          
#> [121] "Sparser_ProtMapper"          "SPIKE"                      
#> [123] "SPIKE_ICELLNET"              "STRING_ICELLNET"            
#> [125] "TCRcuration_SignaLink3"      "TFactS_DoRothEA"            
#> [127] "TFe_DoRothEA"                "TransmiR"                   
#> [129] "TRED_DoRothEA"               "TRIP"                       
#> [131] "TRRD_DoRothEA"               "TRRUST_DoRothEA"            
#> [133] "UniProt_CellPhoneDB"         "UniProt_LRdb"               
#> [135] "Wang"