Retrieve the available resources for a given query type

Collects the names of the resources available in OmniPath for a certain query type and optionally for a dataset within that.

get_resources(query_type, datasets = NULL, generic_categories = NULL)

Arguments

query_type

one of the query types `interactions`, `enz_sub`, `complexes`, `annotations` or `intercell`

datasets

currently within the `interactions` query type only, multiple datasets are available: `omnipath`, `kinaseextra`, `pathwayextra`, `ligrecextra`, `dorothea`, `tf_target`, `tf_mirna`, `mirnatarget` and `lncrna_mrna`.

generic_categories

for the `intercell` query type, restrict the search for some generic categories e.g. `ligand` or `receptor`.

Value

a character vector with resource names

Examples

get_resources(query_type = 'interactions')
#>   [1] "ABS"                         "ACSN"                       
#>   [3] "ACSN_SignaLink3"             "ARACNe-GTEx_DoRothEA"       
#>   [5] "ARN"                         "Adhesome"                   
#>   [7] "AlzPathway"                  "BEL-Large-Corpus_ProtMapper"
#>   [9] "Baccin2019"                  "BioGRID"                    
#>  [11] "CA1"                         "CancerCellMap"              
#>  [13] "CellCall"                    "CellChatDB"                 
#>  [15] "CellChatDB-cofactors"        "CellPhoneDB"                
#>  [17] "CellPhoneDB_Cellinker"       "CellTalkDB"                 
#>  [19] "Cellinker"                   "DEPOD"                      
#>  [21] "DIP"                         "DLRP_Cellinker"             
#>  [23] "DLRP_talklr"                 "DOMINO"                     
#>  [25] "DeathDomain"                 "DoRothEA"                   
#>  [27] "DoRothEA-reviews_DoRothEA"   "ELM"                        
#>  [29] "EMBRACE"                     "ENCODE-distal"              
#>  [31] "ENCODE-proximal"             "ENCODE_tf-mirna"            
#>  [33] "FANTOM4_DoRothEA"            "Fantom5_LRdb"               
#>  [35] "Guide2Pharma"                "Guide2Pharma_CellPhoneDB"   
#>  [37] "Guide2Pharma_Cellinker"      "Guide2Pharma_LRdb"          
#>  [39] "Guide2Pharma_talklr"         "HOCOMOCO_DoRothEA"          
#>  [41] "HPMR"                        "HPMR_Cellinker"             
#>  [43] "HPMR_LRdb"                   "HPMR_talklr"                
#>  [45] "HPRD"                        "HPRD-phos"                  
#>  [47] "HPRD_KEA"                    "HPRD_LRdb"                  
#>  [49] "HPRD_MIMP"                   "HPRD_talklr"                
#>  [51] "HTRIdb"                      "HTRIdb_DoRothEA"            
#>  [53] "HuRI"                        "I2D_CellPhoneDB"            
#>  [55] "ICELLNET"                    "IMEx_CellPhoneDB"           
#>  [57] "InnateDB"                    "InnateDB-All_CellPhoneDB"   
#>  [59] "InnateDB_CellPhoneDB"        "InnateDB_SignaLink3"        
#>  [61] "IntAct"                      "IntAct_CellPhoneDB"         
#>  [63] "IntAct_DoRothEA"             "JASPAR_DoRothEA"            
#>  [65] "KEA"                         "KEGG-MEDICUS"               
#>  [67] "Kinexus_KEA"                 "Kirouac2010"                
#>  [69] "LMPID"                       "LRdb"                       
#>  [71] "Li2012"                      "Lit-BM-17"                  
#>  [73] "LncRNADisease"               "MIMP"                       
#>  [75] "MINT_CellPhoneDB"            "MPPI"                       
#>  [77] "Macrophage"                  "MatrixDB"                   
#>  [79] "MatrixDB_CellPhoneDB"        "NCI-PID_ProtMapper"         
#>  [81] "NFIRegulomeDB_DoRothEA"      "NRF2ome"                    
#>  [83] "NetPath"                     "NetworKIN_KEA"              
#>  [85] "ORegAnno"                    "ORegAnno_DoRothEA"          
#>  [87] "PAZAR"                       "PAZAR_DoRothEA"             
#>  [89] "PhosphoNetworks"             "PhosphoPoint"               
#>  [91] "PhosphoSite"                 "PhosphoSite_KEA"            
#>  [93] "PhosphoSite_MIMP"            "PhosphoSite_ProtMapper"     
#>  [95] "PhosphoSite_noref"           "ProtMapper"                 
#>  [97] "REACH_ProtMapper"            "RLIMS-P_ProtMapper"         
#>  [99] "Ramilowski2015"              "Ramilowski2015_Baccin2019"  
#> [101] "ReMap_DoRothEA"              "Reactome_LRdb"              
#> [103] "Reactome_ProtMapper"         "Reactome_SignaLink3"        
#> [105] "RegNetwork_DoRothEA"         "SIGNOR"                     
#> [107] "SIGNOR_ProtMapper"           "SPIKE"                      
#> [109] "STRING_talklr"               "SignaLink3"                 
#> [111] "Sparser_ProtMapper"          "TCRcuration_SignaLink3"     
#> [113] "TFactS_DoRothEA"             "TFe_DoRothEA"               
#> [115] "TRED_DoRothEA"               "TRIP"                       
#> [117] "TRRD_DoRothEA"               "TRRUST_DoRothEA"            
#> [119] "TransmiR"                    "UniProt_CellPhoneDB"        
#> [121] "UniProt_LRdb"                "Wang"                       
#> [123] "Wojtowicz2020"               "connectomeDB2020"           
#> [125] "dbPTM"                       "iPTMnet"                    
#> [127] "iTALK"                       "miR2Disease"                
#> [129] "miRDeathDB"                  "ncRDeathDB"                 
#> [131] "phosphoELM"                  "phosphoELM_KEA"             
#> [133] "phosphoELM_MIMP"             "scConnect"                  
#> [135] "talklr"