Retrieve the available resources for a given query type

Collects the names of the resources available in OmniPath for a certain query type and optionally for a dataset within that.

Usage

get_resources(query_type, datasets = NULL, generic_categories = NULL)

Arguments

query_type

one of the query types `interactions`, `enz_sub`, `complexes`, `annotations` or `intercell`

datasets

currently within the `interactions` query type only, multiple datasets are available: `omnipath`, `kinaseextra`, `pathwayextra`, `ligrecextra`, `dorothea`, `tf_target`, `tf_mirna`, `mirnatarget` and `lncrna_mrna`.

generic_categories

for the `intercell` query type, restrict the search for some generic categories e.g. `ligand` or `receptor`.

Value

a character vector with resource names

Examples

get_resources(query_type = 'interactions')
#>   [1] "ABS"                         "ACSN"                       
#>   [3] "ACSN_SignaLink3"             "ARACNe-GTEx_DoRothEA"       
#>   [5] "ARN"                         "Adhesome"                   
#>   [7] "AlzPathway"                  "BEL-Large-Corpus_ProtMapper"
#>   [9] "Baccin2019"                  "BioGRID"                    
#>  [11] "CA1"                         "CancerCellMap"              
#>  [13] "CancerDrugsDB"               "CellCall"                   
#>  [15] "CellChatDB"                  "CellChatDB-cofactors"       
#>  [17] "CellPhoneDB"                 "CellPhoneDB_Cellinker"      
#>  [19] "CellTalkDB"                  "Cellinker"                  
#>  [21] "DEPOD"                       "DIP"                        
#>  [23] "DLRP_Cellinker"              "DLRP_talklr"                
#>  [25] "DOMINO"                      "DeathDomain"                
#>  [27] "DoRothEA"                    "DoRothEA-reviews_DoRothEA"  
#>  [29] "ELM"                         "EMBRACE"                    
#>  [31] "ENCODE-distal"               "ENCODE-proximal"            
#>  [33] "ENCODE_tf-mirna"             "FANTOM4_DoRothEA"           
#>  [35] "Fantom5_LRdb"                "Guide2Pharma"               
#>  [37] "Guide2Pharma_CellPhoneDB"    "Guide2Pharma_Cellinker"     
#>  [39] "Guide2Pharma_LRdb"           "Guide2Pharma_talklr"        
#>  [41] "HOCOMOCO_DoRothEA"           "HPMR"                       
#>  [43] "HPMR_Cellinker"              "HPMR_LRdb"                  
#>  [45] "HPMR_talklr"                 "HPRD"                       
#>  [47] "HPRD-phos"                   "HPRD_KEA"                   
#>  [49] "HPRD_LRdb"                   "HPRD_MIMP"                  
#>  [51] "HPRD_talklr"                 "HTRIdb"                     
#>  [53] "HTRIdb_DoRothEA"             "HuRI"                       
#>  [55] "I2D_CellPhoneDB"             "ICELLNET"                   
#>  [57] "IMEx_CellPhoneDB"            "InnateDB"                   
#>  [59] "InnateDB-All_CellPhoneDB"    "InnateDB_CellPhoneDB"       
#>  [61] "InnateDB_SignaLink3"         "IntAct"                     
#>  [63] "IntAct_CellPhoneDB"          "IntAct_DoRothEA"            
#>  [65] "JASPAR_DoRothEA"             "KEA"                        
#>  [67] "KEGG-MEDICUS"                "Kinexus_KEA"                
#>  [69] "Kirouac2010"                 "LMPID"                      
#>  [71] "LRdb"                        "Li2012"                     
#>  [73] "Lit-BM-17"                   "LncRNADisease"              
#>  [75] "MIMP"                        "MINT_CellPhoneDB"           
#>  [77] "MPPI"                        "Macrophage"                 
#>  [79] "MatrixDB"                    "MatrixDB_CellPhoneDB"       
#>  [81] "NCI-PID_ProtMapper"          "NFIRegulomeDB_DoRothEA"     
#>  [83] "NRF2ome"                     "NetPath"                    
#>  [85] "NetworKIN_KEA"               "ORegAnno"                   
#>  [87] "ORegAnno_DoRothEA"           "PAZAR"                      
#>  [89] "PAZAR_DoRothEA"              "PhosphoNetworks"            
#>  [91] "PhosphoPoint"                "PhosphoSite"                
#>  [93] "PhosphoSite_KEA"             "PhosphoSite_MIMP"           
#>  [95] "PhosphoSite_ProtMapper"      "PhosphoSite_noref"          
#>  [97] "ProtMapper"                  "REACH_ProtMapper"           
#>  [99] "RLIMS-P_ProtMapper"          "Ramilowski2015"             
#> [101] "Ramilowski2015_Baccin2019"   "ReMap_DoRothEA"             
#> [103] "Reactome_LRdb"               "Reactome_ProtMapper"        
#> [105] "Reactome_SignaLink3"         "RegNetwork_DoRothEA"        
#> [107] "SIGNOR"                      "SIGNOR_ProtMapper"          
#> [109] "SPIKE"                       "STRING_talklr"              
#> [111] "SignaLink3"                  "Sparser_ProtMapper"         
#> [113] "TCRcuration_SignaLink3"      "TFactS_DoRothEA"            
#> [115] "TFe_DoRothEA"                "TRED_DoRothEA"              
#> [117] "TRIP"                        "TRRD_DoRothEA"              
#> [119] "TRRUST_DoRothEA"             "TransmiR"                   
#> [121] "UniProt_CellPhoneDB"         "UniProt_LRdb"               
#> [123] "Wang"                        "Wojtowicz2020"              
#> [125] "connectomeDB2020"            "dbPTM"                      
#> [127] "iPTMnet"                     "iTALK"                      
#> [129] "lncrnadb"                    "miR2Disease"                
#> [131] "miRDeathDB"                  "miRTarBase"                 
#> [133] "miRecords"                   "ncRDeathDB"                 
#> [135] "phosphoELM"                  "phosphoELM_KEA"             
#> [137] "phosphoELM_MIMP"             "scConnect"                  
#> [139] "talklr"