Imports an intercellular network combining annotations and interactions

Imports an intercellular network by mapping intercellular annotations and protein interactions. First imports a network of protein-protein interactions. Then, it retrieves annotations about the proteins intercellular communication roles, once for the transmitter (delivering information from the expressing cell) and second, the receiver (receiving signal and relaying it towards the expressing cell) side. These 3 queries can be customized by providing parameters in lists which will be passed to the respective methods (import_omnipath_interactions for the network and import_omnipath_intercell for the annotations). Finally the 3 data frames combined in a way that the source proteins in each interaction annotated by the transmitter, and the target proteins by the receiver categories. If undirected interactions present (these are disabled by default) they will be duplicated, i.e. both partners can be both receiver and transmitter. If a cache file provided, its content will be returned without any further filtering.

import_intercell_network(
  cache_file = NULL,
  interactions_param = list(),
  transmitter_param = list(),
  receiver_param = list()
)

Arguments

cache_file
path to an earlier data file; if exists, will be loaded as it is, the further arguments have no effect; if does not exists, the result will be dumped into this file.

interactions_param
a list with arguments for an interactions query: import_omnipath_interactions, import_pathwayextra_interactions, import_kinaseextra_interactions, import_ligrecextra_interactions

transmitter_param
a list with arguments for import_omnipath_intercell, to define the transmitter side of intercellular connections

receiver_param
a list with arguments for import_omnipath_intercell, to define the receiver side of intercellular connections

Value

A dataframe containing information about protein-protein interactions and the inter-cellular roles of the protiens involved in those interactions.

See also

Examples

intercellNetwork <- import_intercell_network(
   interactions_param = list(datasets = 'ligrecextra'),
   receiver_param = list(categories = c('receptor', 'transporter')),
   transmitter_param = list(categories = c('ligand', 'secreted_enzyme')))

#> Downloaded 4074 interactions.

#> Downloaded 10275 intercellular communication role records.

#> Downloaded 23494 intercellular communication role records.