Imports annotations from OmniPath

Imports protein annotations about function, localization, expression, structure and other properties of proteins from OmniPath https://omnipathdb.org/annotations. Note: there might be also a few miRNAs annotated; a vast majority of protein complex annotations are inferred from the annotations of the members: if all members carry the same annotation the complex inherits.

import_omnipath_annotations(
  cache_file = NULL,
  proteins = NULL,
  resources = NULL,
  force_full_download = FALSE,
  wide = FALSE,
  ...
)

import_Omnipath_annotations(...)

import_OmniPath_annotations(...)

Arguments

cache_file
Path to an earlier data file

proteins
Vector containing the genes or proteins for whom annotations will be retrieved (UniProt IDs or HGNC Gene Symbols or miRBase IDs). It is also possible to donwload annotations for protein complexes. To do so, write 'COMPLEX:' right before the genesymbols of the genes integrating the complex. Check the vignette for examples.

resources
Load the annotations only from these databases. See get_annotation_resources for possible values.

force_full_download
Force the download of the entire annotations dataset. This is disabled by default because the size of this data is around 1GB. We recommend to retrieve the annotations for a set of proteins or only from a few resources, depending on your interest.

wide
Convert the annotation table to wide format, which corresponds more or less to the original resource. If the data comes from more than one resource a list of wide tables will be returned.

...
Passed to import_omnipath_annotations.

Value

A data.frame containing different gene/complex annotations

See also

Examples

annotations = import_omnipath_annotations(
    proteins = c('TP53', 'LMNA'),
    resources = c('HPA_subcellular')
)

#> Downloaded 8 annotation records.