Imports OmniPath intercell annotations

Imports the OmniPath intercellular communication role annotation database from https://omnipathdb.org/intercell. It provides information on the roles in inter-cellular signaling. E.g. if a protein is a ligand, a receptor, an extracellular matrix (ECM) component, etc.

import_omnipath_intercell(
  cache_file = NULL,
  categories = NULL,
  resources = NULL,
  parent = NULL,
  scope = NULL,
  aspect = NULL,
  source = NULL,
  transmitter = NULL,
  receiver = NULL,
  secreted = NULL,
  plasma_membrane_peripheral = NULL,
  plasma_membrane_transmembrane = NULL,
  proteins = NULL,
  topology = NULL,
  causality = NULL,
  ...
)

import_Omnipath_intercell(...)

import_OmniPath_intercell(...)

Arguments

cache_file
path to an earlier data file

categories
vector containing the categories to be retrieved. All the genes belonging to those categories will be returned. For further information about the categories see get_intercell_categories

resources
limit the query to certain resources; see the available resources by get_intercell_resources

parent
vector containing the parent classes to be retrieved. All the genes belonging to those classes will be returned. For furter information about the main classes see get_intercell_categories

scope
either `specific` or `generic`

aspect
either `locational` or `functional`

source
either `resource_specific` or `composite`

transmitter
logical, include only transmitters i.e. proteins delivering signal from a cell to its environment

receiver
logical, include only receivers i.e. proteins delivering signal to the cell from its environment

secreted
logical, include only secreted proteins

plasma_membrane_peripheral
logical, include only plasma membrane peripheral membrane proteins

plasma_membrane_transmembrane
logical, include only plasma membrane transmembrane proteins

proteins
limit the query to certain proteins

topology
topology categories: one or more of `secreted` (sec), `plasma_membrane_peripheral` (pmp), `plasma_membrane_transmembrane` (pmtm) (both short or long notation can be used)

causality
`transmitter` (trans), `receiver` (rec) or `both` (both short or long notation can be used)

...
Passed to import_omnipath_intercell.

Value

A dataframe cotaining information about roles in intercellular signaling.

See also

Examples

intercell = import_omnipath_intercell(categories = c('ecm'))

#> Downloaded 2754 intercellular communication role records.