Quality filter for intercell annotations

Quality filter for intercell annotations

intercell_consensus_filter(data, percentile = NULL, loc_percentile = NULL)

Arguments

data
A data frame with intercell annotations, as provided by import_omnipath_intercell.

percentile
Numeric: a percentile cut off for the consensus score of composite categories. The consensus score is the number of resources supporting the classification of an entity into a category based on combined information of many resources. Here you can apply a cut-off, keeping only the annotations supported by a higher number of resources than a certain percentile of each category. If NULL no filtering will be performed. The value is either in the 0-1 range, or will be divided by 100 if greater than 1. The percentiles will be calculated against the generic composite categories and then will be applied to their resource specific annotations and specific child categories.

loc_percentile
Numeric: similar to percentile for major localizations. For example, with a value of 50, the secreted, plasma membrane transmembrane or peripheral attributes will be TRUE only where at least 50 percent of the resources support these.

Value

The data frame in data filtered by the consensus scores.

Examples

intercell <- import_omnipath_intercell(parent = c('ligand', 'receptor'))
nrow(intercell)

#> [1] 50585

# [1] 50174
intercell_q50 <- intercell_consensus_filter(intercell, 50)
nrow(intercell_q50)

#> [1] 43137

# [1] 42863