Interactions from PrePPI

Retrieves predicted protein-protein interactions from the PrePPI database (http://honig.c2b2.columbia.edu/preppi). The interactions in this table are supposed to be correct with a > 0.5 probability.

Usage

preppi_download(...)

Arguments

...

Minimum values for the scores. The available scores are: str, protpep, str_max, red, ort, phy, coexp, go, total, exp and final. Furthermore, an operator can be passed, either .op = '&' or .op = '|', which is then used for combined filtering by multiple scores.

Value

A data frame (tibble) of interactions with scores, databases and literature references.

Details

PrePPI is a combination of many prediction methods, each resulting a score. For an explanation of the scores see https://honiglab.c2b2.columbia.edu/hfpd/help/Manual.html. The minimum, median and maximum values of the scores:

| Score   | Minimum | Median   | Maximum            |
| ------- | ------- | -------- | ------------------ |
| str     |       0 |     5.5  |           6,495    |
| protpep |       0 |     3.53 |          38,138    |
| str_max |       0 |    17.9  |          38,138    |
| red     |       0 |     1.25 |              24.4  |
| ort     |       0 |     0    |           5,000    |
| phy     |       0 |     2.42 |               2.42 |
| coexp   |       0 |     2.77 |              45.3  |
| go      |       0 |     5.86 |             181    |
| total   |       0 | 1,292    | 106,197,000,000    |
| exp     |       1 |   958    |           4,626    |
| final   |     600 | 1,778    |            4.91e14 |

See also

Examples

preppi <- preppi_download()
preppi
#> # A tibble: 1,545,710 × 15
#>    prot1  prot2  str_s…¹ protp…² str_m…³ red_s…⁴ ort_s…⁵ phy_s…⁶ coexp…⁷ go_sc…⁸
#>    <chr>  <chr>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 Q13131 P14625   18.6     6.45   18.6    4.25    0.615    2.42    9.47   10.8 
#>  2 P06400 Q96N96    1.83   14.3    14.3    4.25    0        2.42    2.11   10.8 
#>  3 Q7Z6V5 Q8NCE0    4.57    0       4.57   0       0        1.60    9.47   24.1 
#>  4 P37058 P15428  485.      0     485.     1.77    0.615    2.42    1.18    2.07
#>  5 O00468 Q9NR23   34.0     0      34.0    0.512   0        2.42    2.77    5.86
#>  6 Q9Y2X7 Q9NPJ6    0       2.59    2.59   1.01    0        2.42    1.58    0   
#>  7 Q8IY84 Q9Y3S1  254.      7.35  254.     0.913   0.615    2.42    2.11    2.07
#>  8 Q13315 Q96T68    1.21    9.34    9.34  19.3     0.615    2.42    9.47    0.89
#>  9 Q6GYQ0 Q92529    2.67   20.6    20.6    0       0        2.42    3.90    5.86
#> 10 P08575 O14545    0       0       0      0       0        2.42    3.90  105.  
#> # … with 1,545,700 more rows, 5 more variables: total_score <dbl>, dbs <chr>,
#> #   pubs <chr>, exp_score <dbl>, final_score <dbl>, and abbreviated variable
#> #   names ¹​str_score, ²​protpep_score, ³​str_max_score, ⁴​red_score, ⁵​ort_score,
#> #   ⁶​phy_score, ⁷​coexp_score, ⁸​go_score
# # A tibble: 1,545,710 x 15
#    prot1 prot2 str_score protpep_score str_max_score red_score ort_score
#    <chr> <chr>     <dbl>         <dbl>         <dbl>     <dbl>     <dbl>
#  1 Q131. P146.     18.6           6.45         18.6      4.25      0.615
#  2 P064. Q96N.      1.83         14.3          14.3      4.25      0
#  3 Q7Z6. Q8NC.      4.57          0             4.57     0         0
#  4 P370. P154.    485.            0           485.       1.77      0.615
#  5 O004. Q9NR.     34.0           0            34.0      0.512     0
# # . with 1,545,700 more rows, and 8 more variables: phy_score <dbl>,
# #   coexp_score <dbl>, go_score <dbl>, total_score <dbl>, dbs <chr>,
# #   pubs <chr>, exp_score <dbl>, final_score <dbl>